% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-compensation.R
\name{adaptCompensationSpillmat}
\alias{adaptCompensationSpillmat}
\title{Adapts a spillovermatrix such that it can be used directly for compensation}
\usage{
adaptCompensationSpillmat(input_sm, out_channels)
}
\arguments{
\item{input_sm}{A previously calculated spillover matrix.}

\item{output_names}{The names that the prepared output spillover matrix should have as column names.
Numeric names as well as names of the form MetalMassDi (e.g. Ir191Di or Ir191) will be
interpreted as masses with associated metals.}
}
\value{
An adapted spillovermatrix with column and row names according to the output names
}
\description{
This helper function adapts the columns of a provided spillover matrix such that it is compatible with
data having the column names provided.
}
\details{
The rules how the spillover matrix is adapted can be found in the compCytof function documentation.
}
\examples{
data(ss_exp)

# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)
 re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
spillMat <- computeSpillmat(x = re)
ff <- 
adaptCompensationSpillmat(spillMat, colnames(ss_exp))

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
and Vito Zanotelli \email{vito.zanotelli@uzh.ch}
}
