% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/adaptSpillmat.R
\docType{methods}
\name{adaptSpillmat}
\alias{adaptSpillmat}
\alias{adaptSpillmat,matrix,vector-method}
\title{Adapts a spillovermatrix such that 
it can be used directly for compensation}
\usage{
adaptSpillmat(input_sm, out_chs)

\S4method{adaptSpillmat}{matrix,vector}(input_sm, out_chs)
}
\arguments{
\item{input_sm}{A previously calculated spillover matrix.}

\item{out_chs}{The column names that the prepared output spillover matrix should have.
Numeric names as well as names of the form MetalMass(Di), e.g. Ir191Di 
or Ir191, will be interpreted as masses with associated metals.}

\item{...}{optional arguments.}
}
\value{
An adapted spillover matrix with column and row names 
according to \code{out_chs}.
}
\description{
This helper function adapts the columns of a provided spillover matrix 
such that it is compatible with data having the column names provided.
}
\details{
The rules how the spillover matrix is adapted 
are explained in \link{\code{compCytof}}.
}
\examples{
# get single-stained control samples
data(ss_exp)
# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)
# debarcode
re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
# estimate spillover matrix and adapt it
spillMat <- computeSpillmat(x = re)
adaptCompensationSpillmat(spillMat, colnames(ss_exp))

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
and Vito RT Zanotelli \email{vito.zanotelli@uzh.ch}
}
