% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/applyCutoffs.R
\docType{methods}
\name{applyCutoffs}
\alias{applyCutoffs}
\alias{applyCutoffs,dbFrame-method}
\title{Single-cell debarcoding (2)}
\usage{
applyCutoffs(x, ...)

\S4method{applyCutoffs}{dbFrame}(x, mhl_cutoff = 30)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{further optional arguments.}

\item{mhl_cutoff}{mahalanobis distance threshold above which events should be unassigned.}
}
\value{
Will update the \code{bc_ids} slot of the input \code{\link{dbFrame}}
and return the latter.
}
\description{
Applies separation and mahalanobies distance cutoffs.
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_beads, y = bc_ms)
re <- estCutoffs(x = re)
applyCutoffs(x = re)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}

