% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/assignPrelim.R
\docType{methods}
\name{assignPrelim}
\alias{assignPrelim}
\alias{assignPrelim,flowFrame,data.frame-method}
\alias{assignPrelim,flowFrame,vector-method}
\title{Single-cell debarcoding (1)}
\usage{
assignPrelim(x, y, ...)

\S4method{assignPrelim}{flowFrame,data.frame}(x, y, cofactor = 10,
  verbose = TRUE)

\S4method{assignPrelim}{flowFrame,vector}(x, y, cofactor = 10,
  verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link{flowFrame}}.}

\item{y}{the debarcoding scheme. A binary matrix with sample names as row names and 
numeric masses as column names OR a vector of numeric masses corresponding to 
barcode channels. When the latter is supplied, \code{assignPrelim} will 
create a scheme of the appropriate format internally.}

\item{...}{further optional arguments.}

\item{cofactor}{cofactor used for asinh transformation.}

\item{verbose}{logical. Should extra information on progress be reported? Defaults to TRUE.}
}
\value{
Returns a \code{\link{dbFrame}} containing measured intensities, 
the debarcoding key, a numeric verctor of barcode IDs and separations between 
positive and negative barcode populations, and barcode intensities
normalized by population.
}
\description{
Assigns a preliminary barcode ID to each event.
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
assignPrelim(x = ss_beads, y = bc_ms)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}

