% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concatFCS.R
\docType{methods}
\name{concatFCS}
\alias{concatFCS}
\alias{concatFCS,flowSet-method}
\alias{concatFCS,character-method}
\alias{concatFCS,list-method}
\title{FCS file concatination}
\usage{
concatFCS(x, ...)

\S4method{concatFCS}{flowSet}(x, out_path = NULL)

\S4method{concatFCS}{character}(x, out_path = NULL)

\S4method{concatFCS}{list}(x, out_path = NULL)
}
\arguments{
\item{x}{can be either a \code{\link{flowSet}}, a list of \code{\link{flowFrame}}s, 
a character specifying the location of the FCS files to be concatinated, 
or a vector of FCS file names.}

\item{...}{optional arguments.}

\item{out_path}{an optional character string. If specified, an FCS file 
of the concatinated data will be written to this location. 
If NULL (default), \code{concatFCS} will return a \code{\link{flowFrame}}.}
}
\value{
a \code{\link{flowFrame}} containing measurement intensities 
of all input data or a character of the FCS file name.
}
\description{
Concatinates all FCS files in the specified location
}
\examples{
data(raw_data)
concatFCS(raw_data)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
