% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\alias{daFrame}
\title{Differential analysis frame class}
\usage{
daFrame(fs, panel, md, cols_to_use = NULL, cofactor = 5,
  panel_cols = list(channel = "fcs_colname", antigen = "antigen"),
  md_cols = list(file = "file_name", id = "sample_id", factors =
  c("condition", "patient_id")))
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a data.frame with columns describing the experiment.
An exemplary metadata table could look as follows:\itemize{
\item \code{file_name}: the FCS file name
\item \code{sample_id}: a unique sample identifier
\item \code{condition}: brief sample description (e.g. REF)
\item \code{patient_id}: the patient ID}}

\item{cols_to_use}{a logical vector OR numeric vector of indices 
OR character vector of column names. Specifies the columns to keep 
from the input \code{flowSet}.}

\item{cofactor}{cofactor to use for arcsinh-transformation.}

\item{panel_cols}{a named list specifying column names in the input panel 
that contain i) the channel names of the input \code{flowSet}, and ii) 
the corresponding targeted protein marker. List elements should be named 
\code{"channel"} and \code{"antigen"}, respectively.}

\item{md_cols}{a named list specifying column names in the input metadata
that contain i) the FCS file names, ii) unique sample identifiers, 
iii) a character vector of factors descriptive of the samples
(e.g. condition, treatment, batch, ect.). List elements should be named 
\code{"file"}, \code{"id"}, and \code{"factors"}, respectively.}
}
\description{
Represents the data returned by and used throughout differential analysis.
}
\section{Slots}{

\describe{
\item{\code{assays}}{list of length one containing the arcsinh-transformed expressions.}

\item{\code{rowData}}{the metadata information for each event, and its cluster ID
as inferred by the initial \code{\link{FlowSOM}} clustering.}

\item{\code{colData}}{a data.frame with the following columns:\itemize{
\item \code{channel} original column name in the input \code{flowSet}
\item \code{type1}, \code{type2} logical vectors indicating, 
for each antigen, whether it was used for clustering}}

\item{\code{metadata}}{a named list containing:\itemize{
\item the original metadata-table
\item \code{panel}: the original panel-table
\item \code{n_events}: the number of events measured per sample
\item \code{SOM_codes}: a 100 x p matrix of SOM codes, 
where p = no. of measurement parameters
\item \code{cluster_codes}: cluster codes for the initial 
\code{\link{FlowSOM}} clustering (column \code{'100'}), the 
\code{\link{ConsensusClusterPlus}} metaclustering (columns \code{'2'} 
through {'20'}, and mergings as labeled by \code{\link{mergeClusters}}}}
}}

\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
