% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/daFrame-methods.R
\docType{methods}
\name{n_events}
\alias{n_events}
\alias{type1}
\alias{type2}
\alias{sample_ids}
\alias{cluster_codes}
\alias{cluster_ids}
\alias{exprs,daFrame-method}
\alias{daFrame-methods}
\alias{exprs}
\alias{lineage}
\alias{functional}
\alias{conditions}
\alias{sample_ids}
\alias{cluster_ids}
\alias{merging_codes}
\alias{merging_ids}
\alias{n_events,daFrame-method}
\alias{type1,daFrame-method}
\alias{type2,daFrame-method}
\alias{sample_ids,daFrame-method}
\alias{cluster_ids,daFrame-method}
\alias{cluster_codes,daFrame-method}
\title{Extraction and replacement methods for objects of class \code{daFrame}}
\usage{
n_events(x)

type1(x)

type2(x)

sample_ids(x)

cluster_codes(x)

cluster_ids(x)

\S4method{exprs}{daFrame}(object)

\S4method{n_events}{daFrame}(x)

\S4method{type1}{daFrame}(x)

\S4method{type2}{daFrame}(x)

\S4method{sample_ids}{daFrame}(x)

\S4method{cluster_ids}{daFrame}(x)

\S4method{cluster_codes}{daFrame}(x)
}
\arguments{
\item{x, object}{a \code{\link{daFrame}}.}
}
\value{
\describe{
\item{\code{exprs}}{extracts the arcsinh-transformed expressions.}
\item{\code{n_events}}{extracts the number of events measured per sample.}
\item{\code{type1}}{extracts the antigens used for clustering.}
\item{\code{type2}}{extracts antigens that were not used for clustering.}
\item{\code{sample_ids}}{extracts the sample IDs 
as specified in the metadata-table.}
\item{\code{cluster_ids}}{extracts the numeric vector of cluster IDs
as inferred by \code{\link{FlowSOM}}.}
\item{\code{cluster_codes}}{extracts a \code{data.frame} containing 
cluster codes for the \code{\link{FlowSOM}} clustering, 
the \code{\link{ConsensusClusterPlus}} metaclustering, 
and all mergings done through \code{\link{mergeClusters}}.}
}
}
\description{
Methods for accessing slots in a \code{\link{daFrame}}.
}
\examples{

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
