% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{raw_data}
\alias{sample_ff}
\alias{sample_key}
\alias{ss_exp}
\alias{mp_cells}
\alias{isotope_list}
\title{Example data sets}
\value{
see descriptions above.
}
\description{
\itemize{
\item{Preprocessing
\describe{
\item{\code{raw_data}}{
a \code{\link{flowSet}} with 3 experiments, each containing 2'500 raw 
measurements with a variation of signal over time. Samples were mixed 
with DVS beads capture by mass channels 140, 151, 153, 165 and 175.}
\item{\code{sample_ff}}{
a \code{\link{flowFrame}} following a 6-choose-3 barcoding scheme 
where mass channels 102, 104, 105, 106, 108, and 110 were used for labeling 
such that each of the 20 individual barcodes are positive for exactly 3 
out of the 6 barcode channels.}
\item{\code{sample_key}}{
a \code{data.frame} of dimension 20 x 6 with sample names as row and 
barcode masses as column names. Contains a binary code of length 6 for each
sample in \code{sample_ff}, e.g. 111000, as its unique identifier.}
\item{\code{ss_exp}}{
a \code{\link{flowFrame}} with 20'000 events. 
Contains 36 single-antibody stained controls where beads were stained 
with antibodies captured by mass channels 139, 141 through 156, and 
158 through 176, respectively, and pooled together.}
\item{\code{mp_cells}}{a \code{\link{flowFrame}} with 5000 spill-affected
multiplexed cells and 39 measurement parameters.}
\item{\code{isotope_list}}{a named list of isotopic compositions for all 
elements within 75 through 209 u corresponding to the  CyTOF mass range
at the time of writing (see reference).}}}
\item{
}
}
}
\examples{
### example data for normalization:
    # raw measurement data
    data(raw_data)
  
### example data for debarcoding:
    # 20 barcoded samples
    data(sample_ff)
    # 6-choose-3 barcoding scheme
    data(sample_key)

### example data for compensation:
    # single-stained control samples
    data(ss_exp)
    # multiplexed cells
    data(mp_cells)

}
\references{
Coursey, J.S., Schwab, D.J., Tsai, J.J., Dragoset, R.A. (2015).
Atomic weights and isotopic compositions, 
(available at http://physics.nist.gov/Comp).
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
