% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R, R/dbFrame-methods.R
\docType{class}
\name{dbFrame-class}
\alias{bc_ids}
\alias{bc_ids,dbFrame-method}
\alias{bc_ids-method}
\alias{bc_key}
\alias{bc_key,dbFrame-method}
\alias{bc_key-method}
\alias{counts}
\alias{counts,dbFrame-method}
\alias{counts-method}
\alias{dbFrame}
\alias{dbFrame-class}
\alias{dbFrame-methods}
\alias{deltas}
\alias{deltas,dbFrame-method}
\alias{deltas-method}
\alias{exprs}
\alias{exprs,dbFrame-method}
\alias{exprs-method}
\alias{mhl_cutoff}
\alias{mhl_cutoff,dbFrame-method}
\alias{mhl_cutoff-method}
\alias{normed_bcs}
\alias{normed_bcs,dbFrame-method}
\alias{normed_bcs-method}
\alias{sep_cutoffs}
\alias{sep_cutoffs,dbFrame-method}
\alias{sep_cutoffs-method}
\alias{yields}
\alias{yields,dbFrame-method}
\alias{yields-method}
\title{Debarcoding frame class}
\usage{
bc_key(object)

bc_ids(object)

deltas(object)

normed_bcs(object)

sep_cutoffs(object)

mhl_cutoff(object)

counts(object)

yields(object)

\S4method{exprs}{dbFrame}(object)

\S4method{bc_key}{dbFrame}(object)

\S4method{bc_ids}{dbFrame}(object)

\S4method{deltas}{dbFrame}(object)

\S4method{normed_bcs}{dbFrame}(object)

\S4method{sep_cutoffs}{dbFrame}(object)

\S4method{mhl_cutoff}{dbFrame}(object)

\S4method{counts}{dbFrame}(object)

\S4method{yields}{dbFrame}(object)
}
\arguments{
\item{object}{a \code{dbFrame}.}
}
\description{
Class representing the data returned by and used throughout debarcoding.
}
\details{
\code{show(dbFrame)} will display \itemize{
\item the dimensionality of the measurement data and number of barcodes used
\item current assignments in order of decreasing population size
\item separation cutoff estimates
\item the average and per-population yield achieven upon debarcoding}
}
\section{Slots}{

\describe{
\item{\code{exprs}}{a matrix containing raw intensities of the input flowFrame.}

\item{\code{bc_key}}{binary barcoding scheme with numeric masses as column names OR 
a numeric vector specifying indices of columns corresponding to barcode channels.}

\item{\code{bc_ids}}{numeric vector of barcode IDs. If a barcoding scheme is supplied, the respective 
binary code's row index, else, the column index of the respective barcode channel.}

\item{\code{deltas}}{numeric vector of separations between positive and negative barcode populations 
computed from normalized barcode intensities.}

\item{\code{normed_bcs}}{matrix containing normalized barcode intensities.}

\item{\code{sep_cutoffs}}{numeric vector of distance separation cutoffs between positive and negative 
barcode populations above which events will be unassigned.}

\item{\code{mhl_cutoff}}{non-negative numeric value specifying the Mahalanobis distance cutoffs
below which events will be unassigned.}

\item{\code{counts}}{matrix of dimension (# barcodes)x(101) where each row contains
the number of events within a barcode for which positive and negative populations 
are separated by a distance between in [0,0.01), ..., [0.99,1], respectively.}

\item{\code{yields}}{a matrix of dimension (# barcodes)x(101) where each row contains
the percentage of events within a barcode that will be obtained after applying
a separation cutoff of 0, 0.01, ..., 1, respectively.}
}}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

