% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/debarcode.R
\docType{methods}
\name{debarcode}
\alias{debarcode}
\alias{debarcode,flowFrame,data.frame-method}
\alias{debarcode,flowFrame,vector-method}
\title{Compute debarcoding}
\usage{
debarcode(x, y, out_path, ...)

\S4method{debarcode}{flowFrame,data.frame}(x, y, out_path, cofactor = 10,
  mhl_cutoff = 30, verbose = TRUE)

\S4method{debarcode}{flowFrame,vector}(x, y, out_path, cofactor = 10,
  mhl_cutoff = 30, verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link{flowFrame}}.}

\item{y}{the debarcoding scheme. A binary matrix with sample names as row names and 
numeric masses as column names OR a vector of numeric masses corresponding to 
barcode channels. When the latter is supplied, \code{assignPrelim} will 
create a scheme of the appropriate format internally.}

\item{out_path}{a character string. All outputs (population-wise FCS files and 
diagnostic plots) will be generated in this location.}

\item{...}{further optional arguments.}

\item{cofactor}{cofactor used for asinh transformation.}

\item{mhl_cutoff}{mahalanobis distance threshold above which events should be unassigned.}

\item{verbose}{logical. Should extra information on progress be reported? Defaults to TRUE.}
}
\value{
Will generate population-wise FCS files, and events and yields plots
in the specified \code{out_path}.
}
\description{
For each barcode, estimates a cutoff parameter for the 
distance between positive and negative barcode populations.
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_beads, y = bc_ms)
estCutoffs(x = re)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

