% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffAbundance.R
\docType{methods}
\name{diffAbundance}
\alias{diffAbundance}
\alias{diffAbundance,daFrame-method}
\title{Differential population abundances}
\usage{
diffAbundance(x, k, K, ...)

\S4method{diffAbundance}{daFrame}(x, k, K)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{k}{specifies which clustering to perform the analysis on.}

\item{K}{contrasts.}

\item{...}{optional arguments.}
}
\value{
Writes a table with frequencies across samples,
non-adjusted and adjuster p-values for each cluster into the
\code{metadata} slot of the input \code{daFrame}.
}
\description{
Compares the proportions of cell types across experimental conditions 
and aims to highlight populations that are present at different ratios.
}
\details{

}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
# specify contrasts
K <- matrix(c(0, 1), nrow=1, byrow=TRUE, dimnames=list("BCRXLvsRef"))
re <- differentialAbundance(re, 12, K)

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
