% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/differentialAbundance.R
\docType{methods}
\name{differentialAbundance}
\alias{differentialAbundance}
\alias{differentialAbundance,flowSet,data.frame,numeric,matrix-method}
\title{differentialAbundance}
\usage{
differentialAbundance(fs, md, cluster_ids, K, ...)

\S4method{differentialAbundance}{flowSet,data.frame,numeric,matrix}(fs, md,
  cluster_ids, K)
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a data.frame containing the metadata.}

\item{cluster_ids}{a numeric vector.}

\item{K}{contrasts.}

\item{...}{optional arguments.}
}
\value{
Returns a table with frequencies across samples,
non-adjusted and adjuster p-values for each cluster.
}
\description{
Compares the proportions of cell types across experimental conditions 
and aims to highlight populations that are present at different ratios.
}
\details{

}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
