% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/extractClusters.R
\docType{methods}
\name{extractClusters}
\alias{extractClusters}
\alias{extractClusters,daFrame,character-method}
\title{Extract clusters from a \code{daFrame}}
\usage{
extractClusters(x, k, ...)

\S4method{extractClusters}{daFrame,character}(x, k, clusters = NULL,
  as = c("flowSet", "fcs"), out_dir = ".", verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{k}{numeric or character string. 
Specifies the clustering to extract populations from.
Must be one of \code{names(cluster_codes(x))}.}

\item{...}{optional arguments.}

\item{clusters}{a character vector.
Specifies which clusters to extract.
\code{NULL} = all clusters.}

\item{as}{\code{"flowSet"} or \code{"fcs"}. 
Specifies whether clusters should be return 
as a \code{flowSet} or written to FCS files.}

\item{out_dir}{a character string.
Specifies where FCS files should be writen to.
Defaults to the working directory.}

\item{verbose}{logical.
Should information on progress be reported?}
}
\value{
a \code{flowSet} or character vector of the output file names.
}
\description{
Extracts clusters from a \code{daFrame}. Populations will be 
either returned as a \code{flowSet} or written to FCS files,
depending on argument \code{as}.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
re <- cluster(re)

# merge clusters
re <- mergeClusters(re, k="meta20", table=merging_table, id="merging_1")
extractClusters(re, k="merging_1")

}
\author{
Mark D Robinson,
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
