% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2sce.R
\name{fcs2sce}
\alias{fcs2sce}
\title{`SingleCellExperiment` from FCS file(s)}
\usage{
fcs2sce(x, by_time = TRUE, file_no = FALSE)
}
\arguments{
\item{x}{character vector of FCS file names.}

\item{by_time}{logical. Should files be ordered by acquisition time?}

\item{file_no}{logical. Should a file number parameter be added?}
}
\description{
Concatenates input FCS files to a `SingleCellExperiment`.
}
\author{
Helena L. Crowell
}
