% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCytof.R
\name{gateCytof}
\alias{gateCytof}
\title{Gating on CyTOF data}
\usage{
gateCytof(
  x,
  chs,
  type = c("rect", "elip", "live"),
  xy = c(1, 1),
  q = 0.99,
  k = NULL,
  i = 1,
  s = 0.5,
  gate_id = NULL,
  group_by = NULL,
  gate_on = NULL,
  assay = "exprs",
  overwrite = FALSE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{chs}{character string of length 2 specifying the channels to gate on. 
Should be one of \code{rownames(x)} or a \code{[int_]colData(x)} column,
and correspond to a numeric variable.}

\item{type}{character string specifying the gate type:
"rect" for rectangular, "elip" for eliptical, or "live" for polygonal.}

\item{q}{numeric in (0,1) giving the quantile(s) 
for gate of type "elip" and "live".}

\item{k}{integer number of cluster centers for eliptical gates.}

\item{i, s}{numeric specifying line intercept & slope for "live" gates.}

\item{gate_id}{character string giving a unique identifier for the gate.}

\item{group_by}{character string specifying the grouping variable
in case gates should be applied to specified cell subsets (e.g., 
indenpendently for each sample). Should be a \code{[int_]colData(x)} 
column, and correspond to a logical or factor variable.}

\item{gate_on}{character string specifying a previously applied gate
to gate on, i.e., from which to subset cells prior to gating.
Should be one of \code{names(int_metadata(x)$gates)}.}

\item{assay}{character string specifying which assay data to use.
Should be one of \code{assayNames(x)}.}

\item{overwrite}{logical. If there already exists a gate 
with identifier \code{gate_id}, should it be overwritten?}
}
\value{
a \code{SingleCellExperiment} with gating information stored in
\code{int_metadata(.)$gates$<gate_id>} as a list containing:
\describe{
\item{\code{chs}}{channels that were gated on (length 2 character vector)}
\item{\code{type}}{gating type (one of "rect", "elip", "live")}
\item{\code{pars}}{list of applied gating parameters
(e.g., group-specific quantiles for "elip"/"live" gates)}
\item{\code{gate_on}}{gate ID of an upstream gate determining a cell subset}
\item{\code{group_by}}{variable by which cells were grouped by}
\item{\code{data}}{\code{data.frame} storing the gate's metadata 
(e.g., x- and y-range for rectangular gates; 
gate boundaries for "elip" and "live" gates)}
}
}
\description{
Wrapper around \code{openCyto} gating methods for 
\code{SingleCellExperiment}s, including group-specific gating parameters.
}
\examples{
data(raw_data)
sce <- fcs2sce(raw_data)

# GATING FOR CELLS

# specify DNA channels
dna_chs <- c("Ir191Di", "Ir193Di")

# apply eliptical gate
sce <- gateCytof(sce, dna_chs, 
  group_by = "file_name", gate_id = "cells",
  type = "elip", q = 0.95, xy = c(4, 4))
  
# view number & fraction of selected cells
table(sce$cells)
mean(sce$cells)

# view gating metadata
library(SingleCellExperiment)
gi <- int_metadata(sce)$gates
head(gi$cells$data) # data.frame of gate boundaries
gi$cells$pars       # list of applied parameters

# visualize gate on scatter plot
plotScatter(sce, gate_id = "cells")

}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
