% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchGUI.R
\name{launchGUI}
\alias{launchGUI}
\title{Launch GUI}
\usage{
launchGUI()
}
\value{
Opens a browser window with an interactive Shiny application.
}
\description{
Launches the CATALYST Shiny app.
}
\details{
Detailed user guides are available inside the app. To use the app 
online, please visit \emph{http://imlspenticton.uzh.ch:3838/CATALYSTLite}.
}
\examples{
\dontrun{launchGUI}

}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
