% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mergeClusters.R
\docType{methods}
\name{mergeClusters}
\alias{mergeClusters}
\alias{mergeClusters,daFrame-method}
\title{Manual cluster merging}
\usage{
mergeClusters(x, table, label)

\S4method{mergeClusters}{daFrame}(x, table, label)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{table}{the merging table; a data.frame with columns 
\code{'old_cluster'}, \code{'new_cluster'} and \code{'label'}.}

\item{label}{a character string.}

\item{...}{optional arguments.}
}
\value{
Writes the newly assigend cluster codes into the metadata slot 
\code{cluster_codes} of the input \code{daFrame} and returns the latter.
}
\description{

}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
re <- mergeClusters(re, merging_table, "merging_1")
plotClusterHeatmap(re, k=20, m="merging_1", functional="CD4")

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
