% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCytof.R
\name{normCytof}
\alias{normCytof}
\title{Bead-based normalization}
\usage{
normCytof(
  x,
  beads,
  assay = "exprs",
  k = 500,
  trim = 5,
  remove_beads = TRUE,
  norm_to = NULL,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{beads}{\code{"dvs"} (for bead masses 140, 151, 153 ,165, 175) or 
\code{"beta"} (for bead masses 139, 141, 159, 169, 175) or 
a numeric vector of masses.}

\item{assay}{character string specifying which assay data to use.
Should be one of \code{assayNames(x)}.}

\item{k}{integer width of the median window used 
for bead smoothing (affects visualizations only!).}

\item{trim}{a single non-negative numeric. 
A \emph{median}+/-\code{trim}*\emph{mad} rule is applied to 
preliminary bead populations to remove bead-bead doublets and 
low signal beads prior to estimating normalization factors.}

\item{remove_beads}{logical. If TRUE, bead events will be removed from
the input \code{SingleCellExperiment} and returned as a separate object?}

\item{norm_to}{a \code{\link[flowCore]{flowFrame}} or path to
an FCS file from which to compute baseline bead intensities,
and to which the input data should be normalized to.}

\item{plot}{logical. Should bead vs. DNA scatters and smoothed bead 
intensities before vs. after normalization be included in the output?}

\item{verbose}{logical. Should extra information on progress be reported?}
}
\value{
a list of the following \code{SingleCellExperiment}... 
\itemize{
\item{\code{data}: SCE with additional assay slot \code{normed} containing 
  normalized (and optionally filtered, if \code{remove_beads = TRUE}) data}
\item{\code{beads}, \code{removed}: SCEs containing subsets of events 
  identified as beads and that were removed, respectively. 
  The latter includes bead-cell and cell-cell doublets)}
} ...and \code{ggplot} objects:
\itemize{
\item{\code{scatter}: scatter plot of DNA vs. bead 
  intensities with indication of the applied gates}
\item{\code{lines}: running-median smoothed bead 
  intensities before and after normalization}
}
}
\description{
an implementation of Finck et al.'s normalization of mass cytometry data 
using bead standards with automated bead gating.
}
\examples{
data(raw_data)
sce <- fcs2sce(raw_data)
res <- normCytof(sce, beads = "dvs", k = 80)

ncol(res$beads)   # no. of bead events
ncol(res$removed) # no. of events removed

res$scatter # plot DNA vs. bead intensities including applied gates
res$lines   # plot smoothed bead intensities before vs. after normalization

}
\references{
Finck, R. et al. (2013). 
Normalization of mass cytometry data with bead standards.
\emph{Cytometry A} \bold{83A}, 483-494.
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
