% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outFCS.R
\docType{methods}
\name{outFCS}
\alias{outFCS}
\alias{outFCS,dbFrame,character-method}
\title{Write population-wise FCS}
\usage{
outFCS(x, out_path)

\S4method{outFCS}{dbFrame,character}(x, out_path)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{out_path}{character string. Specifies in which location 
output files are to be generated.}

\item{...}{further optional arguments.}
}
\description{
Writes an FCS file for each barcode population.
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_beads, y = bc_ms)
#outFCS(x = re, y = ...)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

