% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outFrames.R
\docType{methods}
\name{outFrames}
\alias{outFrames}
\alias{outFrames,dbFrame-method}
\title{Population-wise \code{flowFrame}s from a \code{dbFrame}}
\usage{
outFrames(x, ...)

\S4method{outFrames}{dbFrame}(x, return = "flowSet", which = "assigned")
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{return}{\code{"flowSet"} or \code{"list"}. Specifies the output type.}

\item{which}{Specifies which barcode(s) to include. \code{"assigned"} (if the population
of unassigned events should be excluded), \code{"all"} (if the latter should
be included), or a numeric or character specifying a subset of populations. 
Valid values are IDs that occur as row names in the \code{bc_key} of the 
supplied \code{\link{dbFrame}}. Defaults to \code{"assigned"}.}
}
\value{
a \code{\link{flowSet}} or list of \code{\link{flowFrame}}s.
}
\description{
Returns a \code{flowSet} or list of \code{flowFrame}s from a 
\code{\link{dbFrame}}. Each \code{flowFrame} will contain the subset of 
events that have been assigned to the same ID.
}
\details{
Creates a separate \code{\link{flowFrame}} for each barcode population 
and, if desired, the population of unassigned events.
}
\examples{
data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
outFrames(x = re, return = "list", which = c("B1", "D4"))

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
