% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotAbundances.R
\docType{methods}
\name{plotAbundances}
\alias{plotAbundances}
\alias{plotAbundances,daFrame-method}
\title{Population frequencies across samples & clusters}
\usage{
plotAbundances(x, ...)

\S4method{plotAbundances}{daFrame}(x, k = 20, by = c("sample_id",
  "cluster_id"))
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{k}{specifies which clustering to use.}

\item{by}{a character string specifying whether 
to plot frequencies across samples or clusters.}
}
\value{
a \code{ggplot} object.
}
\description{

}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
# plot relative population abundances 
plotAbundances(re, k=12)                 # ...across samples 
plotAbundances(re, k=8, by="cluster_id") # ...across clusters

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
