% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotClusterExprs.R
\docType{methods}
\name{plotClusterExprs}
\alias{plotClusterExprs}
\alias{plotClusterExprs,daFrame-method}
\title{Plot expression distributions by cluster}
\usage{
plotClusterExprs(x, ...)

\S4method{plotClusterExprs}{daFrame}(x, k = 20, markers = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{k}{numeric or character string. Specifies the clustering to use.}

\item{markers}{character string specifying which markers to include. Defaults to NULL 
(= all markers). Alternatively, if the \code{colData(x)$marker_class} 
column is specified, can be one of "type", "state", or "none".}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plots smoothed densities of arcsinh-transformed 
marker intensities by cluster.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotClusterExprs(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
