% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotClusterHeatmap.R
\docType{methods}
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\alias{plotClusterHeatmap,daFrame-method}
\title{Median marker expressions across clusters}
\usage{
plotClusterHeatmap(x, ...)

\S4method{plotClusterHeatmap}{daFrame}(x, hm2 = NULL, k = 20, m = NULL,
  freq_bars = TRUE, freq_labs = TRUE, scaled = TRUE, out_path = NULL)
}
\arguments{
\item{x}{expression matrix.}

\item{...}{optional arguments.}

\item{hm2}{a character string that specifies the right-hand side heatmap. 
Depending on the argument, the output will be as follows:
\itemize{
\item \code{"abundances"}: cluster frequencies across samples
\item \code{"type2"}: median type 2 marker expressions across clusters 
(analogous to the left-hand side heatmap)
\item a character string/vector corresponding to one/multiple marker(s): 
median marker expressions across samples and clusters}}

\item{k}{specifies the clustering across which 
median marker expressions should be computed.}

\item{m}{specifies the second layer of clustering to be shown.}

\item{freq_bars, freq_labs}{logical. Should marginal histograms
of cluster frequencies be shown and annotated?}

\item{scaled}{logical specifying whether scaled values should be plotted
(see below for details).}

\item{out_path}{a character string. If specified, 
output will be generated in this location. Defaults to NULL.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Plots a heatmap of median lineage marker expressions across clusters, 
and median functional marker expressions within each cluster across samples.
}
\details{
Scaled values corresponds to cofactor arcsinh-transformed 
expression values scaled between 0 and 1 using 1% and 99% percentiles as 
boundaries. Hierarchical clustering is performed on the unscaled data.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33", 
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

plotClusterHeatmap(re, hm2="abundances")
plotClusterHeatmap(re, hm2="type2", k=12)
plotClusterHeatmap(re, hm2="pS6", k=12, m=8)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
