% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCounts.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,daFrame-method}
\title{Plot cell counts}
\usage{
plotCounts(x, ...)

\S4method{plotCounts}{daFrame}(x, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{color_by}{character string. Must appear as a column name of \code{rowData(x)}. 
Specifies the color coding.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Barplot of the number of cells measured for each sample.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotCounts(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
