% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDeltaArea.R
\docType{methods}
\name{plotDeltaArea}
\alias{plotDeltaArea}
\alias{plotDeltaArea,list-method}
\title{Relative change in are under CDF curve}
\usage{
plotDeltaArea(x, ...)

\S4method{plotDeltaArea}{list}(x)
}
\arguments{
\item{...}{optional arguments.}

\item{fs}{}

\item{md}{}
}
\value{

}
\description{

}
\details{
The delta area represents the amount of extra cluster stability gained when 
clustering into k groups as compared to k-1 groups. It can be expected that 
high stability of clusters can be reached when clustering into the number of 
groups that best fits the data. The ???natural??? number of clusters present in 
the data should thus corresponds to the value of k where there is no longer 
a considerable increase in stability (pleateau onset).
}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
