% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDiffHeatmap.R
\docType{methods}
\name{plotDiffHeatmap}
\alias{plotDiffHeatmap}
\alias{plotDiffHeatmap,daFrame-method}
\title{Plot differential heatmap}
\usage{
plotDiffHeatmap(x, ...)

\S4method{plotDiffHeatmap}{daFrame}(x, y, type = c("DA", "DS"), top_n = 20,
  all = FALSE, order = TRUE, th = 0.1)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{y}{a \code{SummarizedExperiment} containing differential testing 
results as returned by one of \code{\link{testDA_edgeR}}, \code{\link{testDA_voom}}, 
\code{\link{testDA_GLMM}}, \code{\link{testDS_limma}}, or \code{\link{testDS_LMM}}.}

\item{type}{character string. 
Specifies whether to plot a heatmap for differential abundance (DA) 
or differential state (DS) test results.}

\item{top_n}{numeric. 
Number of top clusters (if \code{type = "DA"}) or cluster-marker 
combinations (if \code{type = "DS"}) to display. Defaults to 20.}

\item{all}{logical. 
Specifies whether to display all clusters or cluster-marker combinations. 
If \code{TRUE}, \code{top_n} will be ignored.}

\item{order}{logical.
Specifies whether results should be ordered by adjusted p-value.}

\item{th}{numeric.
Threshold on adjusted p-values below which clusters (DA) or
cluster-marker combinations (DS) should be considered significant.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Heatmaps summarizing differental abundance 
& differential state testing results.
}
\details{
For DA tests, \code{plotDiffHeatmap} will display
\itemize{
\item median (arcsinh-transformed) cell-type marker expressions (across all samples)
\item cluster abundances by samples
\item row annotations indicating the significance of detected clusters
}
For DS tests, \code{plotDiffHeatmap} will display
\itemize{
\item median (arcsinh-transformed) cell-type marker expressions (across all samples)
\item median (arcsinh-transformed) cell-state marker expressions (across all samples)
\item median (arcsinh-transformed) cell-state marker expressions by sample
\item row annotations indicating the significance of detected cluster-marker combinations
}
}
\author{
Lukas M Weber, Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
