% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDR.R
\docType{methods}
\name{plotDR}
\alias{plotDR}
\alias{plotDR,daFrame-method}
\title{Plot dim. reduction from a \code{daFrame}}
\usage{
plotDR(x, ...)

\S4method{plotDR}{daFrame}(x, dr = c("TSNE", "PCA", "MDS", "UMAP",
  "DiffusionMap"), color_by = "meta20", facet = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{dr}{character string specifying the dimensionaly reduction method.}

\item{color_by}{character string specifying a clustering,
marker, or \code{rowData} column name to color by.}

\item{facet}{a character string specifying 
a \code{rowData} column to facet by.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot cell-level reduced dimensions
  stored within a \code{daFrame} object.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
daf <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

daf <- runDR(daf, "PCA")
plotDR(daf, "PCA", color_by = "condition")

daf <- cluster(daf)
plotDR(daf, "PCA", color_by = "meta5")

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch}
}
