% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvents.R
\name{plotEvents}
\alias{plotEvents}
\title{Event plot}
\usage{
plotEvents(
  x,
  which = "all",
  assay = "scaled",
  n = 1000,
  out_path = NULL,
  name_ext = NULL
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{which}{\code{"all"}, numeric or character. Specifies which barcode(s) to plot. 
Valid values are IDs that occur as row names in the \code{bc_key} of the 
supplied \code{\link{dbFrame}}, or 0 for unassigned events.}

\item{assay}{character string specifying which 
assay data slot to use. One of \code{assayNames(x)}.}

\item{n}{single numeric specifying the number of events to plot.}

\item{out_path}{character string. If specified, outputs will be generated here.}

\item{name_ext}{character string. If specified, will be appended to the file name.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
Plots normalized barcode intensities for a given barcode.
}
\details{
Plots intensities normalized by population for each barcode specified
by \code{which}: Each event corresponds to the intensities plotted on a 
vertical line at a given point along the x-axis. Events are scaled to the 
95\% quantile of the population it has been assigned to. Barcodes with 
less than 50 event assignments will be skipped; it is strongly recoomended
to remove such populations or reconsider their separation cutoffs.
}
\examples{
data(sample_ff, sample_key)
es <- as.matrix(exprs(sample_ff))
library(SingleCellExperiment)
sce <- SingleCellExperiment(
    assays = list(counts = t(es)),
    rowData = pData(parameters(sample_ff)))
sce <- assignPrelim(x = sce, bc_key = sample_key)
plotEvents(sce, out_path = "~/Desktop")

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena L. Crowell
}
