% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExprHeatmap.R
\name{plotExprHeatmap}
\alias{plotExprHeatmap}
\title{Plot expression heatmap}
\usage{
plotExprHeatmap(
  x,
  features = NULL,
  scale = TRUE,
  row_anno = TRUE,
  row_clust = TRUE,
  col_clust = TRUE,
  row_dend = TRUE,
  col_dend = TRUE,
  draw_freqs = FALSE,
  bin_anno = FALSE,
  hm_pal = brewer.pal(9, "YlGnBu"),
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  linkage = c("average", "ward.D", "single", "complete", "mcquitty", "median",
    "centroid", "ward.D2")
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{features}{character string specifying which features to include;
valid values are \code{"type"/"state"} for \code{type/state_markers(x)} 
if \code{rowData(x)$marker_class} have been specified; 
a subset of \code{rownames(x)}; NULL to use all features.}

\item{scale}{logical specifying whether expressions should be scaled
using lower (1\%) and upper (99\%) quantiles as boundaries;
hierarchical clustering is performed on unscaled data, 
regardless of whether \code{scale = TRUE} or \code{FALSE}.}

\item{row_anno}{logical. Should row annotations for 
each non-numeric cell metadata factor be included?}

\item{row_clust, col_clust}{logical specifying 
whether rows/columns (samples/features) should be 
hierarchically clustered and re-ordered accordingly.}

\item{row_dend, col_dend}{logical specifying whether to include the
row/column dendrogram for the hierarchical clustering of samples/markers.}

\item{draw_freqs}{logical specifying whether to display
a barplot of cell counts labeled with proportions 
for each sample as a right-hand side row annotation.}

\item{bin_anno}{logical. Specifies whether to display values inside bins.}

\item{hm_pal}{character vector of colors to interpolate for the heatmap.}

\item{distance}{character string specifying the distance metric 
to use in \code{\link[stats]{dist}} for hierarchical clustering.}

\item{linkage}{character string specifying the agglomeration method 
to use in \code{\link[stats]{hclust}} for hierarchical clustering.}
}
\value{
a \code{\link[ComplexHeatmap]{Heatmap-class}} object.
}
\description{
Heatmap of median marker expressions by sample, include annotation 
of cell metadata factors as well as relative and absolute cell counts.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)

# turn everything on
plotExprHeatmap(sce, 
  bin_anno = TRUE,
  draw_freqs = TRUE)
  
# turn everything off
plotExprHeatmap(sce,
  row_anno = FALSE,
  row_dend = FALSE,
  col_dend = FALSE)

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
