% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExprs.R
\name{plotExprs}
\alias{plotExprs}
\title{Plot expressions}
\usage{
plotExprs(x, features = NULL, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{features}{specifies which features to include.
Either a character string specifying a subset of features,
or NULL for all features. When \code{rowData(x)$marker_class} 
is specified, can be one of "type", "state", or "none".}

\item{color_by}{character string corresponding to a 
\code{colData(x)} column. Specifies the color coding.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plots the smoothed densities of arcsinh-transformed marker intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
plotExprs(sce)

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
