% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,daFrame-method}
\title{Median marker expressions}
\usage{
plotHeatmap(x, ...)

\S4method{plotHeatmap}{daFrame}(x, clustering = NULL, out_path = NULL)
}
\arguments{
\item{x}{expression matrix.}

\item{...}{optional arguments.}

\item{clustering}{a numeric vector of length 2. 
Specifies the number of clusters across which median marker expressions 
should be computed and the second layer of clustering to be shown. 
If NULL (the default), medians will be computed across samples.}

\item{out_path}{a character string. If specified, 
output will be generated in this location. Defaults to NULL.}
}
\value{


Heatmap of median marker expressions
}
\description{

}
\details{
Displayed values corresponds to cofactor 5 arcsinh-transformed 
expression values scaled between 0 and 1 using 1% and 99% percentiles as 
boundaries. Hierarchical clustering is performed on the unscaled data.
}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
