% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMDS.R
\docType{methods}
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,flowSet,data.frame,data.frame-method}
\title{MDS plot}
\usage{
plotMDS(fs, panel, md, ...)

\S4method{plotMDS}{flowSet,data.frame,data.frame}(fs, panel, md,
  cols = c("steelblue", "indianred"))
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a \code{data.frame} containing the following columns:
\itemize{
\item \code{file_name}: names of the fcs files contained in the input flowSet
\item \code{sample_id}: a unique identifier for each sample
\item \code{condition}: sample description (e.g. healthy or diseased)
}}

\item{...}{optional arguments.}
}
\value{
Plots a multi-dimensional scaling (MDS) plot 
on median marker expression values.
}
\description{
Multi-dimensional scaling (MDS) plot on median marker expressions.
}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
