% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMedExprs.R
\name{plotMedExprs}
\alias{plotMedExprs}
\title{Median expressions}
\usage{
plotMedExprs(
  x,
  k = "meta20",
  features = "state",
  facet_by = c("antigen", "cluster_id"),
  group_by = "condition",
  shape_by = NULL
)
}
\arguments{
\item{x}{a \code{\link{SingleCellExperiment}{SingleCellExperiment}}.}

\item{k}{character string specifying which clustering to use;
values values are \code{names(cluster_codes(x))}.
Ignored if \code{facet_by = "antigen"}.}

\item{features}{character vector specifying 
which features to include; valid values are 
\code{"type"/"state"} for \code{type/state_markers(x)} 
if \code{rowData(x)$marker_class} have been specified; 
a subset of \code{rownames(x)}; NULL to use all features.}

\item{facet_by}{\code{"antigen"} or \code{"cluster_id"}; 
the latter requires having run \code{\link{cluster}}.}

\item{group_by}{character string specifying a non-numeric cell metadata 
variable to group samples by; valid values are \code{names(colData(x))}.}

\item{shape_by}{character string specifying a non-numeric cell metadata 
variable to shape samples by; valid values are \code{names(colData(x))}.}
}
\value{
a \code{ggplot} object.
}
\description{
Boxplots of median marker expressions by sample
and colored by a non-numeric cell metadata variable of interest.
}
\examples{
# construct SCE
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)

# plot median expressions
plotMedExprs(sce, 
  shape_by = "patient_id",
  features = sample(rownames(sce), 6))

# run clustering
sce <- cluster(sce)

# plot median expressions across clusters
p <- plotMedExprs(sce, facet_by = "cluster_id", k = "meta8")
p$facet$params$ncol <- 4; p

# change colors & facetting layout
library(ggplot2)
p$facet$params$ncol <- 4
p + scale_color_manual(values = c(
  Ref = "royalblue", BCRXL = "orange"))

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
