% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNRS.R
\docType{methods}
\name{plotNRS}
\alias{plotNRS}
\alias{plotNRS,daFrame-method}
\title{Plot non-redundancy scores}
\usage{
plotNRS(x, ...)

\S4method{plotNRS}{daFrame}(x, markers = NULL, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{markers}{character string specifying which markers to include. Defaults to NULL 
(= all markers). Alternatively, if the \code{colData(x)$marker_class} 
column is specified, can be one of "type", "state", or "none".}

\item{color_by}{character string. Has to appeara as a column name of \code{rowData(x)}.
Specifies the color coding.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots non-redundancy scores (NRS) by markers in decreasing order.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotNRS(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
