% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotProportions.R
\docType{methods}
\name{plotProportions}
\alias{plotProportions}
\alias{plotProportions,flowSet,data.frame,numeric-method}
\title{Population frequencies across samples & clusters}
\usage{
plotProportions(fs, md, cluster_ids, ...)

\S4method{plotProportions}{flowSet,data.frame,numeric}(fs, md, cluster_ids,
  by = c("sample", "cluster"))
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a \code{data.frame} containing the following columns: \itemize{
\item \code{file_name}: names of the fcs files contained in the input flowSet
\item \code{sample_id}: a unique identifier for each sample
\item \code{condition}: sample description (e.g. healthy or diseased)}}

\item{cluster_ids}{numeric vector of cluster ids for each event in \code{fs}}

\item{...}{optional arguments.}

\item{by}{character specifying whether to plot relative abundances 
across samples or clusters}
}
\value{
Barplot of relative population abundances.
}
\description{

}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
