% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotScatter.R
\docType{methods}
\name{plotScatter}
\alias{plotScatter}
\alias{plotScatter,dbFrame-method}
\title{Channel-wise medians of measurement vs. compensated data}
\usage{
plotScatter(x, ...)

\S4method{plotScatter}{dbFrame}(x, SM, out_path = NULL, name_ext = NULL)
}
\arguments{
\item{x}{a \code{dbFrame}.}

\item{...}{optional arguments.}

\item{out_path}{a character string. If specified, outputs will be generated in this location. 
Defaults to NULL.}

\item{name_ext}{a character string. If specified, will be appended to the plot's name. Defaults to NULL.}

\item{CM}{matrix returned from \code{computeCompmat}.}
}
\description{
For each barcode populations, shows the channel-wise medians
computed from measurement and compensated data.
}
\examples{
data(ss_exp)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
spillMat <- computeSpillmat(x = re)
plotScatter(x = re, SM = spillMat)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

