% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSNE.R
\docType{methods}
\name{plotSNE}
\alias{plotSNE}
\alias{plotSNE,daFrame-method}
\title{MDS plot}
\usage{
plotSNE(x, ...)

\S4method{plotSNE}{daFrame}(x, color = 20, facette = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{color}{numeric value between 2 and 20 OR a character string specifying
an antibody that appears in the metadata table of the input \code{daFrame}.
Specifies the color coding.}

\item{facette}{one of \code{NULL}, \code{"sample_id"} or \code{"condition"}.
Specifies whether the data should be split between sample IDs or conditions,
respectively. Defaults to NULL.}
}
\value{
a \code{ggplot} object.
}
\description{
Multi-dimensional scaling (MDS) plot on median marker expressions.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
re <- tSNE(re)
plotSNE(re, "CD4")

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
