% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_ff}
\alias{sample_ff}
\title{20 sample experiment}
\format{A flowFrame with 10000 cells and 64 observables.}
\usage{
data(sample_ff)
}
\description{
A flowFrame obtained from a triple-staining experiment.
}
\details{
Following a 6-choose-3 barcoding scheme, mass channels 
102, 104, 105, 106, 108, and 110 where used for labeling
such that each of the 20 individual barcodes are positive 
for exactly 3 of the 6 possible barcode channels.
}
\keyword{datasets}

