% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_key}
\alias{sample_key}
\title{6-choose-3 barcoding scheme}
\format{A data.frame with 20 rows (samples) and 6 columns (barcodes).}
\usage{
data(sample_key)
}
\description{
A data.frame with dimensions 20 x 6.
}
\details{
A 6-choose-3 barcoding scheme, with mass channels 102, 104, 105, 106, 108, 
and 110 used for labeling. Each of the 20 individual barcodes are positive 
for exactly 3 of the 6 possible barcode channels. For each sample, 
the barcoding scheme contains a binary code of length 6, e.g. 111000, 
as a unique identifier.
}
\keyword{datasets}

