% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCE-utils.R
\docType{methods}
\name{ei}
\alias{ei}
\alias{n_cells}
\alias{marker_classes}
\alias{type_markers}
\alias{state_markers}
\alias{sample_ids}
\alias{cluster_ids}
\alias{cluster_codes}
\alias{ei,SingleCellExperiment-method}
\alias{exprs}
\alias{n_cells,SingleCellExperiment-method}
\alias{marker_classes,SingleCellExperiment-method}
\alias{type_markers,SingleCellExperiment-method}
\alias{state_markers,SingleCellExperiment-method}
\alias{cluster_codes,SingleCellExperiment-method}
\alias{sample_ids,SingleCellExperiment-method}
\alias{cluster_ids,SingleCellExperiment,missing-method}
\alias{cluster_ids,SingleCellExperiment,character-method}
\title{\code{\link{SingleCellExperiment}} convencience functions}
\usage{
ei(x)

n_cells(x)

marker_classes(x)

type_markers(x)

state_markers(x)

sample_ids(x)

cluster_ids(x, k)

cluster_codes(x)

\S4method{ei}{SingleCellExperiment}(x)

\S4method{n_cells}{SingleCellExperiment}(x)

\S4method{marker_classes}{SingleCellExperiment}(x)

\S4method{type_markers}{SingleCellExperiment}(x)

\S4method{state_markers}{SingleCellExperiment}(x)

\S4method{cluster_codes}{SingleCellExperiment}(x)

\S4method{sample_ids}{SingleCellExperiment}(x)

\S4method{cluster_ids}{SingleCellExperiment,missing}(x, k = NULL)

\S4method{cluster_ids}{SingleCellExperiment,character}(x, k = NULL)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{k}{character string specifying the clustering to extract.
Valid values are \code{names(cluster_codes(x))}.}
}
\value{
\describe{
\item{\code{ei}}{
extracts the experimental design table.}
\item{\code{n_cells}}{
extracts the number of events measured per sample.}
\item{\code{marker_classes}}{
extracts marker class assignments ("type", "state", "none").}
\item{\code{type_markers}}{
extracts the antigens used for clustering.}
\item{\code{state_markers}}{
extracts antigens that were not used for clustering.}
\item{\code{sample_ids}}{
extracts the sample IDs as specified in the metadata-table.}
\item{\code{cluster_codes}}{
extracts a \code{data.frame} containing cluster codes for the 
\code{\link{FlowSOM}} clustering, the \code{\link{ConsensusClusterPlus}} 
metaclustering, and all mergings done through \code{\link{mergeClusters}}.}
\item{\code{cluster_ids}}{
extracts the numeric vector of cluster IDs 
as inferred by \code{\link{FlowSOM}}.}
}
}
\description{
Various wrappers to conviniently access slots
  in a \code{\link[SingleCellExperiment]{SingleCellExperiment}}
  created with \code{\link{prepData}}, and that are used 
  frequently during differential analysis.
}
\examples{
# construct SCE & run clustering
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
sce <- cluster(sce)

# view experimental design table
ei(sce)

# quick-access sample & cluster assignments
plot(table(sample_ids(sce)))
plot(table(cluster_ids(sce)))

# access marker information
marker_classes(sce)
type_markers(sce)
state_markers(sce)

# get cluster ID correspondece between 2 clusterings
old_ids <- seq_len(20)
m <- match(old_ids, cluster_codes(sce)$`meta20`)
new_ids <- cluster_codes(sce)$`meta12`[m]
data.frame(old_ids, new_ids)

# plot relative change in area under CDF curve vs. k
SingleCellExperiment::metadata(sce)$delta_area
}
\author{
Helena L Crowell
}
