% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_exp}
\alias{ss_exp}
\title{Single-stained beads}
\format{A flowFrame with 10000 cells and 61 observables.}
\usage{
data(ss_exp)
}
\description{
A flowFrame obtained from a 36ab-panel single-staining experiment.
}
\details{
36ab-panel: Beads were stained for antibodies captured by 
mass channels 139, 141 through 156, and 158 through 176, respectively, 
and pooled together. Here we sampled a fraction of all recorded events 
to demonstrate the debarcoding and compensation work-flow, 
at the cost of not necessarily arriving at satisfying results.
}
\keyword{datasets}

