% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCytof.R
\name{transform}
\alias{transform}
\title{...}
\usage{
transform(
  x,
  method = "arcsinh",
  cf = 10,
  assay = "counts",
  id = "exprs",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{method}{character or function.}

\item{cf}{numeric cofactor for arcsinh-transformation.}

\item{assay}{character. Valid values are \code{assayNames}.}

\item{id}{character. ID to be given to newly generated assay slot.}

\item{overwrite}{logical. Should assay slot \code{id} be overwritten?}

\item{...}{optional arguments passed to \code{method}.}
}
\description{
...
}
\examples{
data(raw_data)
sce <- fcs2sce(raw_data)

}
\author{
Helena L. Crowell
}
