% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyCutoffs.R
\name{applyCutoffs}
\alias{applyCutoffs}
\title{Single-cell debarcoding (2)}
\usage{
applyCutoffs(x, assay = "exprs", mhl_cutoff = 30, sep_cutoffs = NULL)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{mhl_cutoff}{mahalanobis distance threshold above which events should be unassigned. 
This argument will be ignored if the \code{mhl_cutoff} slot of the input 
\code{dbFrame} is specified.}

\item{sep_cutoffs}{non-negative numeric of length one or of same length as the number of rows 
in the \code{bc_key(x)}. Specifies the distance separation cutoffs between 
positive and negative barcode populations below which events should be 
unassigned. If \code{NULL} (default), \code{applyCutoffs} will try to 
access the \code{sep_cutoffs} slot of the input \code{dbFrame}.}
}
\value{
the input \code{SingleCellExperiment} \code{x} is returned with 
updated \code{colData} columns \code{"bc_id"} and \code{"mhl_dist"}, 
and an additional \code{int_metadata} slot \code{"mhl_cutoff"} 
containing the applied mahalanobies distance cutoff.
}
\description{
Applies separation and mahalanobies distance cutoffs.
}
\examples{
library(SingleCellExperiment)

# construct SCE
data(sample_ff, sample_key)
es <- as.matrix(exprs(sample_ff))
sce <- SingleCellExperiment(
    assays = list(counts = t(es)),
    rowData = pData(parameters(sample_ff)))
assay(sce, "exprs") <- asinh(assay(sce, "counts") / 10)
    
# assign preliminary barcode IDs
# & estimate separation cutoffs
sce <- assignPrelim(x = sce, bc_key = sample_key)
sce <- estCutoffs(sce)

# apply cutoffs
sce <- estCutoffs(x = sce)
sce <- applyCutoffs(x = sce, sep_cutoffs = 0)
plotEvents(sce, "A1", out_path = "~/Desktop", n = 1e3)

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena L. Crowell
}
