% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPrelim.R
\name{assignPrelim}
\alias{assignPrelim}
\title{Single-cell debarcoding (1)}
\usage{
assignPrelim(x, y, cofactor = 10, verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{y}{the debarcoding scheme. A binary matrix with sample names as row
names and numeric masses as column names OR a vector of numeric masses 
corresponding to barcode channels. When the latter is supplied, 
`assignPrelim` will create a scheme of the appropriate format internally.}

\item{cofactor}{numeric. Cofactor used for asinh transformation.}

\item{verbose}{logical. Should extra information on progress be reported?}
}
\value{
a \code{SingleCellExperiment} structured as follows: 
\describe{
\item{assays}{
\itemize{
\item{\code{counts} - raw counts}
\item{\code{exprs} - arcsinh-transformed counts}
\item{\code{scaled} - population-wise scaled 
expression using (95\%)-quantiles as boundaries}
}
}
\item{\code{colData}}{
\itemize{
\item{bc_id}{numeric verctor of barcode assignments}
\item{delta}{separation between positive and negative barcode populations}
}
}
\item{\code{metadata}}{
\itemize{
\item{bc_key}{the input debarcoding scheme}
}
}
}
}
\description{
Assigns a preliminary barcode ID to each event.
}
\examples{
data(sample_ff, sample_key)
es <- as.matrix(exprs(sample_ff))
sce <- SingleCellExperiment(
    assays = list(counts = t(es)),
    rowData = pData(parameters(sample_ff)))
assignPrelim(x = sce, y = sample_key)

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena L. Crowell
}
