% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cluster.R
\docType{methods}
\name{cluster}
\alias{cluster}
\alias{cluster,daFrame-method}
\title{FlowSOM clustering and ConsensusClusterPlus metaclustering}
\usage{
cluster(x, ...)

\S4method{cluster}{daFrame}(x, cols_to_use, facette = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{cols_to_use}{a character vector.
Specifies which antigens to use for clustering.}

\item{facette}{one of \code{NULL}, \code{"sample_id"} or \code{"condition"}.
Specifies whether the data should be split between sample IDs or conditions,
respectively. Defaults to NULL.}
}
\value{
a \code{ggplot} object.
}
\description{
\code{CATALYST::cluster()} runs \code{\link{FlowSOM}} clustering into 100,
and \code{\link{ConsensusClusterPlus}} metaclustering into 2-20 clusters. 
In the returned \code{daFrame}, those antigens used for clustering will be 
labelled as 'type1', and the remainder of antigens as 'type2'. A binary 
indication of each marker's type can be viewed via \code{colData(daFrame)}. 
Differential analysis should be performed on type2 markers exclusively.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# specify antigens to use for clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33", 
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
(re <- cluster(re, cols_to_use=lineage))

# sanity check
all.equal(lineage, type1(re))

# get type2 markers for differential analysis
type2(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
