% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/compCytof.R
\docType{methods}
\name{compCytof}
\alias{compCytof}
\alias{compCytof,flowFrame,matrix-method}
\title{Compensate CyTOF experiment}
\usage{
compCytof(x, y, ...)

\S4method{compCytof}{flowFrame,matrix}(x, y, out_path = NULL)
}
\arguments{
\item{x}{a \code{\link{flowFrame}} OR a character string specifying 
the location of FCS files that should be compensates.}

\item{y}{a spillover matrix.}

\item{...}{further optional arguments.}

\item{out_path}{a character string. If specified, compensated FCS files will be generated 
in this location. If \code{x} is a character string, file names will be 
inherited from uncompensated FCS files and given extension "_comped".
Defaults to NULL.}
}
\value{
Compensates the input \code{\link{flowFrame}} OR, 
if \code{x} is a character string, all FCS files in the specified location.
}
\description{
For each barcode, estimates a cutoff parameter for the 
distance between positive and negative barcode populations.
}
\examples{
data(ss_exp)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
spillMat <- computeSpillmat(x = re)
compCytof(ss_exp, spillMat)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
