% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/computeSpillmat.R
\docType{methods}
\name{computeSpillmat}
\alias{computeSpillmat}
\alias{computeSpillmat,dbFrame-method}
\title{Compute spillover matrx}
\usage{
computeSpillmat(x, ...)

\S4method{computeSpillmat}{dbFrame}(x, method = "mean", trim = 0.08)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{method}{function to be used for computing spillover estimates. 
Defaults to \code{mean}.}

\item{trim}{trim value used for estimation of spill values. Note that
\code{trim = 0.5} is equivalent to \code{method = "median"}.}
}
\value{
Returns a square compensation matrix with dimensions and dimension names 
matching those of the input flowFrame. Spillover is assumed to be linear and 
is thence computed as the ratio of a positive barcode population's median 
or (trimmed) mean intensity in affected and spilling channel. Furthermore, 
on the basis of their additive nature, spillover values are computed 
independently for each interacting pair of channels. The current framework 
considers only potential (not all possible) interactions, that is,
M+/-1 (detection sensitivity), same metals (isotopic impurites) and M+16M 
(oxide formation). By default, diagonal entries are set to 1.
}
\description{
Computes the spillover matrix based on 
a priori identified single-positive popultions.
}
\examples{
# get single-stained control samples
data(ss_exp)

# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)

re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
head(computeSpillmat(x = re))

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
