% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concatFCS.R
\docType{methods}
\name{concatFCS}
\alias{concatFCS}
\alias{concatFCS,flowSet-method}
\title{FCS file concatination}
\usage{
concatFCS(x, ...)

\S4method{concatFCS}{flowSet}(x, out_path = NULL)
}
\arguments{
\item{x}{can be either a \code{\link{flowSet}}, a list of \code{\link{flowFrame}}s, 
a character specifying the location of the FCS files to be concatinated, 
or a vector of FCS file names.}

\item{...}{optional arguments.}

\item{out_path}{an optional character string. If specified, an FCS file 
of the concatinated data will be written to this location. 
If NULL (default), \code{concatFCS} will return a \code{\link{flowFrame}}.}
}
\value{
Returns either a \code{\link{flowFrame}} or FCS files containing 
the measured intensities of all provided FCS files. Files will be 
linked together in the order of appearance in the specified location.
}
\description{
Concatinates all FCS files in the specified location
}
\examples{
path <- system.file("extdata", package="CATALYST")
fcsFiles <- list.files(path, "data_\\\\d+.fcs", full.names=TRUE)
concatFCS(fcsFiles)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
