% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concatFCS.R
\docType{methods}
\name{concatFCS}
\alias{concatFCS}
\alias{concatFCS,flowSet-method}
\alias{concatFCS,character-method}
\alias{concatFCS,list-method}
\title{FCS file Concatenation}
\usage{
concatFCS(x, ...)

\S4method{concatFCS}{flowSet}(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE, pars = NULL, desc = NULL)

\S4method{concatFCS}{character}(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE)

\S4method{concatFCS}{list}(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE)
}
\arguments{
\item{x}{can be either a \code{flowSet}, a list of \code{flowFrame}s, 
a character specifying the location of the FCS files to be concatinated, 
or a vector of FCS file names.}

\item{...}{optional arguments.}

\item{out_path}{character string. If specified, an FCS file of the concatenated data 
will be written to this location. 
If NULL (default), a \code{flowFrame} will be returned.}

\item{by_time}{logical. Specifies whether files should be ordered by time of acquisition.}

\item{file_num}{logical. Specifies whether a file number column should be added.}

\item{pars, desc}{optional character vectors of channel names & descriptions
to use when merging files.}
}
\value{
a \code{flowFrame} containing measurement intensities 
of all input data or a character of the FCS file name.
}
\description{
Concatenates all input data to a single file or object.
}
\examples{
data(raw_data)
concatFCS(raw_data)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
