% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concatFCS.R
\docType{methods}
\name{concatFCS}
\alias{concatFCS}
\alias{concatFCS,character-method}
\title{FCS file concatination}
\usage{
concatFCS(x, ...)

\S4method{concatFCS}{character}(x, out_path = NULL)
}
\arguments{
\item{x}{a character specifying the location of the FCS files to be concatinated
or a vector of FCS file names}

\item{...}{optional arguments.}

\item{out_path}{an optional character string. If specified, an FCS file 
of the concatinated data will be written to this location. 
If NULL (default), \code{concatFCS} will return a \code{\link{flowFrame}}.}
}
\value{
Returns either a \code{\link{flowFrame}} or FCS files containing 
the measured intensities of all provided FCS files. Files will be 
linked together in the order of appearance in the specified location.
}
\description{
Concatinates all FCS files in the specified location
}
\examples{

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
