% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\title{Differential analysis frame class}
\description{
Represents the data returned by 
  and used throughout differential analysis.

\describe{
\item{\code{assays}}{a list of length 1 containing the measurement data.}
\item{\code{rowData}}{
  an object of the \code{\link[S4Vectors]{DataFrame-class}}
  containing the row metadata.}
\item{\code{colData}}{
  an object of the \code{\link[S4Vectors]{DataFrame-class}}
  containing the column metadata.}
\item{\code{metadata}}{
  a named list containing: \itemize{
  \item{\code{experimental_design}: the original metadata table.}
  \item{\code{n_cells}: number of events (cells) measured per sample.}
  \item{\code{cofactor}: numeric cofactor used for arcsinh-transformation.}
  \item{\code{cluster_codes}: 
  cluster codes for the initial \pkg{FlowSOM} clustering (\code{"som100"}), 
  the \pkg{ConsensusClusterPlus} metaclustering (\code{"metaX"}), and all 
  manual mergings done with \code{\link{mergeClusters}}.}
  \item{\code{delta_area}: the Delta Area plot 
  (see \pkg{ConsensusClusterPlus} for details).}
  \item{\code{experimental_design}: the original metadata table}}}}
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
