% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\alias{dbFrame}
\title{Differential analysis frame class}
\description{
Represents the data returned by and used throughout differential analysis.
}
\details{
Objects of class \code{daFrame} 
hold all data required for differential analysis:
}
\section{Slots}{

\describe{
\item{\code{data}}{a \link{\code{flowSet}} holding all samples.}

\item{\code{panel}}{}

\item{\code{cluster_ids}}{}

\item{\code{merging_ids}}{}
}}

\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
