% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\alias{daFrame}
\title{Differential analysis frame class}
\usage{
daFrame(x, panel, md, cols_to_use = NULL, cofactor = 5,
  panel_cols = list(channel = "fcs_colname", antigen = "antigen", class =
  "marker_class"), md_cols = list(file = "file_name", id = "sample_id",
  factors = c("condition", "patient_id")))
}
\arguments{
\item{x}{a \code{flowSet} holding all samples or a path to a set of FCS files.}

\item{panel}{a data.frame containing, for each channel, 
its column name in the input data, targeted protein marker,
and (optionally) class ("type", "state", or "none").}

\item{md}{a table with column describing the experiment.
An exemplary metadata table could look as follows: \itemize{
\item\code{file_name}: the FCS file name
\item\code{sample_id}: a unique sample identifier
\item\code{patient_id}: the patient ID
\item\code{condition}: brief sample description 
  (e.g. reference/stimulated, healthy/diseased)}}

\item{cols_to_use}{a logical vector, numeric vector of column indices,
or character vector of channel names. Specified which column to keep 
from the input data. Defaults to the channels listed in the input panel.}

\item{cofactor}{numeric cofactor to use for arcsinh-transformation.}

\item{panel_cols}{a named list specifying the column names of \code{md}
that contain the FCS file names, sample IDs, and factors of interest
(batch, condition, treatment etc.).}

\item{md_cols}{a names list specifying the column names of \code{panel}
that contain the channel names, targeted protein markers, and (optionally) 
marker classes.}
}
\value{
an object of class \code{daFrame}.
}
\description{
Represents the data returned by and used throughout differential analysis.

\describe{
\item{\code{assays}}{a list of length 1 containing the measurement data.}
\item{\code{rowData}}{a \code{\link{DataFrame}} containing the row metadata.}
\item{\code{colData}}{a \code{\link{DataFrame}} containing the column metadata.}
\item{\code{metadata}}{a named list containing: \itemize{
  \item{\code{experimental_design}: the original metadata table.}
  \item{\code{n_cells}: number of events (cells) measured per sample.}
  \item{\code{cofactor}: numeric cofactor used for arcsinh-transformation.}
  \item{\code{cluster_codes}:
    cluster codes for the initial \pkg{FlowSOM} clustering (\code{"som100"}), 
    the \pkg{ConsensusClusterPlus} metaclustering (\code{"metaX"}), and all
    manual mergings done with \code{\link{mergeClusters}}.}
  \item{\code{delta_area}: the Delta Area plot 
    (see \pkg{ConsensusClusterPlus} for details).}
  \item{\code{experimental_design}: the original metadata table}}}}
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
daFrame(PBMC_fs, PBMC_panel, PBMC_md)

}
\author{
Helena Lucia Crowell \email{helena@crowells.eu}
}
