% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\alias{daFrame}
\title{Differential analysis frame class}
\usage{
daFrame(fs, panel, md, cols_to_use = NULL, cofactor = 5,
  panel_cols = list(channel = "fcs_colname", antigen = "antigen"),
  md_cols = list(file = "file_name", id = "sample_id", factors =
  c("condition", "patient_id")))
}
\arguments{
\item{fs}{a \code{\link[flowCore]{flowSet}} holding all samples.}

\item{panel}{a 2 column \code{data.frame} that contains for each marker of interest 
i) its column name in the FCS file, and ii) the targeted protein marker.}

\item{md}{a \code{data.frame} with columns describing the experiment.
An exemplary metadata table could look as follows:
\itemize{
  \item \code{file_name}: the FCS file name
  \item \code{sample_id}: a unique sample identifier 
  \item \code{condition}: brief sample description (e.g. REF) 
  \item \code{patient_id}: the patient ID}}

\item{cols_to_use}{a logical vector OR numeric vector of indices OR character vector 
of column names. Specifies the columns to keep from \code{fs}.}

\item{cofactor}{numeric. Cofactor to use for arcsinh-transformation.}

\item{panel_cols}{a named list specifying column names of \code{panel} that contain i) the 
original channel names in \code{fs}, and ii) the targeted protein marker. 
Elements must be named \code{"channel"} and \code{"antigen"}.}

\item{md_cols}{a named list specifying column names of \code{md} that contain i) the FCS 
file names, ii) unique sample identifiers, and iii) a character vector of 
factors descriptive of the samples (e.g. condition, treatment, ect.).
Elements must be named \code{"file"}, \code{"id"}, and \code{"factors"}.}
}
\value{
an object of class \code{\link{SummarizedExperiment}}.
}
\description{
Represents the data returned by and used throughout differential analysis.
}
\section{Slots}{

\describe{
\item{\code{assays}}{list of length one containing the arcsinh-transformed expressions.}

\item{\code{rowData}}{the metadata information for each event, and its cluster ID
as inferred by the initial \code{\link{FlowSOM}} clustering.}

\item{\code{colData}}{a data.frame with the following columns:\itemize{
\item \code{marker_name} original column name in the input \code{flowSet}
\item \code{marker_class} one of \code{"type"} or \code{"state"}}}

\item{\code{metadata}}{a named list containing:\itemize{
\item \code{design}: the original metadata-table
\item \code{panel}: the original panel-table
\item \code{n_cells}: the number of events measured per sample
\item \code{SOM_codes}: a k x p matrix of SOM codes, where
  k = no. of clusters, and p = no. of measurement parameters
\item \code{cluster_codes}: cluster codes for the initial \pkg{FlowSOM} 
  clustering, the \pkg{ConsensusClusterPlus} metaclustering, and manual 
  mergings done with \code{\link{mergeClusters}}}}
}}

\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
