% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{daFrame-class}
\alias{daFrame-class}
\alias{daFrame}
\title{Differential analysis frame class}
\usage{
daFrame(fs, panel, md, cols_to_use, cofactor = 5)
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{panel}{a data.frame with the following columns:
\itemize{
\item \code{file_name}: the FCS file name
\item \code{sample_id}: a unique sample identifier
\item \code{condition}: brief sample description (e.g. REF)
\item \code{patient_id}: the patient ID}}

\item{md}{a data.frame with the following columns:
\itemize{
\item \code{fcs_colname}: the marker's column name in the FCS file
\item \code{Antigen}: the targeted protein marker}}

\item{cols_to_use}{a logical vector OR numeric vector of indices OR 
character vector. Specifies columns to keep from the input \code{flowSet}.}

\item{cofactor}{cofactor to use for arcsinh-transformation.}
}
\description{
Represents the data returned by and used throughout differential analysis.
}
\section{Slots}{

\describe{
\item{\code{assays}}{list of length one containing the arcsinh-transformed expressions.}

\item{\code{rowData}}{the metadata information for each event, and its cluster ID
as inferred by the initial \code{\link{FlowSOM}} clustering.}

\item{\code{colData}}{a data.frame with the following columns:\itemize{
\item \code{channel} original column name in the input \code{flowSet}
\item \code{type1}, \code{type2} logical vectors indicating, 
for each antigen, whether it was used for clustering}}

\item{\code{metadata}}{a named list containing:\itemize{
\item the original metadata-table
\item \code{panel}: the original panel-table
\item \code{n_events}: the number of events measured per sample
\item \code{SOM_codes}: a 100 x p matrix of SOM codes, 
where p = no. of measurement parameters
\item \code{cluster_codes}: cluster codes for the initial 
\code{\link{FlowSOM}} clustering (column \code{'100'}), the 
\code{\link{ConsensusClusterPlus}} metaclustering (columns \code{'2'} 
through {'20'}, and mergings as labeled by \code{\link{mergeClusters}}}}
}}

\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
