% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/daFrame-methods.R
\docType{methods}
\name{daFrame-methods}
\alias{daFrame-methods}
\alias{panel}
\alias{metadata}
\alias{cluster_ids}
\alias{merging_ids}
\alias{data,daFrame-method}
\alias{daFrame-methods}
\alias{data}
\alias{panel}
\alias{metadata}
\alias{cluster_ids}
\alias{merging_ids}
\alias{cluster_ids<-}
\alias{merging_ids<-}
\alias{panel,daFrame-method}
\alias{metadata,daFrame-method}
\alias{cluster_ids,daFrame-method}
\alias{merging_ids,daFrame-method}
\title{Extraction and replacement methods for objects of class \code{daFrame}}
\usage{
data(x)

panel(x)

metadata(x)

cluster_ids(x)

merging_ids(x)

\S4method{data}{daFrame}(x)

\S4method{panel}{daFrame}(x)

\S4method{metadata}{daFrame}(x)

\S4method{cluster_ids}{daFrame}(x)

\S4method{merging_ids}{daFrame}(x)
}
\arguments{
\item{x, object}{a \code{\link{daFrame}}.}

\item{value}{the replacement value.}
}
\value{
\describe{
\item{\code{data}}{extracts the raw data.}
\item{\code{panel}}{extracts the panel.}
\item{\code{metadata}}{extracts the metadata.}
\item{\code{cluster_ids}}{extracts current cluster assignments.}
\item{\code{merging_ids}}{extracts second layer of cluster assignments.}
}
}
\description{
Methods for replacing and accessing slots in a \code{\link{daFrame}}.
}
\examples{

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
