% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/daFrame-methods.R,
%   R/daFrame-subsetting.R
\docType{methods}
\name{ei}
\alias{ei}
\alias{n_cells}
\alias{marker_classes}
\alias{type_markers}
\alias{state_markers}
\alias{sample_ids}
\alias{cluster_codes}
\alias{cluster_ids}
\alias{reducedDimNames,daFrame-method}
\alias{daFrame-methods}
\alias{exprs}
\alias{reducedDims,daFrame-method}
\alias{reducedDim,daFrame-method}
\alias{reducedDimNames<-,daFrame,character-method}
\alias{exprs,daFrame-method}
\alias{ei,daFrame-method}
\alias{n_cells,daFrame-method}
\alias{marker_classes,daFrame-method}
\alias{type_markers,daFrame-method}
\alias{state_markers,daFrame-method}
\alias{sample_ids,daFrame-method}
\alias{cluster_codes,daFrame-method}
\alias{cluster_ids,daFrame-method}
\alias{[,daFrame,ANY,ANY,ANY-method}
\title{Extraction and replacement methods 
for objects of class \code{daFrame}}
\usage{
ei(x)

n_cells(x)

marker_classes(x)

type_markers(x)

state_markers(x)

sample_ids(x)

cluster_codes(x)

cluster_ids(x, ...)

\S4method{reducedDimNames}{daFrame}(x)

\S4method{reducedDims}{daFrame}(x)

\S4method{reducedDim}{daFrame}(x, dr = 1)

\S4method{reducedDimNames}{daFrame,character}(x) <- value

\S4method{exprs}{daFrame}(object)

\S4method{ei}{daFrame}(x)

\S4method{n_cells}{daFrame}(x)

\S4method{marker_classes}{daFrame}(x)

\S4method{type_markers}{daFrame}(x)

\S4method{state_markers}{daFrame}(x)

\S4method{sample_ids}{daFrame}(x)

\S4method{cluster_codes}{daFrame}(x)

\S4method{cluster_ids}{daFrame}(x, k = NULL)

\S4method{[}{daFrame,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x, object}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{dr}{character string specifying the dim. reduction to extract.}

\item{value}{a character vector containing 
the desired dimensionality reduction names.}

\item{k}{character string specifying the clustering to extract.
Valid values are \code{names(cluster_codes(x))}.}

\item{i, j}{indices specifying elements to extract.}
}
\value{
\describe{
\item{\code{exprs}}{
extracts the arcsinh-transformed expressions.}
\item{\code{n_cells}}{
extracts the number of events measured per sample.}
\item{\code{type_markers}}{
extracts the antigens used for clustering.}
\item{\code{state_markers}}{
extracts antigens that were not used for clustering.}
\item{\code{sample_ids}}{
extracts the sample IDs as specified in the metadata-table.}
\item{\code{cluster_codes}}{
extracts a \code{data.frame} containing cluster codes for the 
\code{\link{FlowSOM}} clustering, the \code{\link{ConsensusClusterPlus}} 
metaclustering, and all mergings done through \code{\link{mergeClusters}}.}
\item{\code{cluster_ids}}{
extracts the numeric vector of cluster IDs 
as inferred by \code{\link{FlowSOM}}.}
}
}
\description{
Methods for accessing slots in a \code{\link{daFrame}}.
}
\examples{
# construct daFrame
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
re <- cluster(re)

# view data summary
library(SummarizedExperiment)
cbind(metadata(re)$experiment_info, cells=n_cells(re))

# access row / cell data
head(rowData(re))
plot(table(cluster_ids(re)))

# access marker information
type_markers(re)
state_markers(re)

# get cluster ID correspondece between 2 clusterings
old_ids <- seq_len(20)
m <- match(old_ids, cluster_codes(re)$`meta20`)
new_ids <- cluster_codes(re)$`meta12`[m]
data.frame(old_ids, new_ids)

# plot relative change in area under CDF curve vs. k
metadata(re)$delta_area
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
