% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/daFrame-methods.R
\docType{methods}
\name{daFrame-methods}
\alias{daFrame-methods}
\alias{functional}
\alias{sample_ids}
\alias{cluster_codes}
\alias{cluster_ids}
\alias{exprs,daFrame-method}
\alias{daFrame-methods}
\alias{exprs}
\alias{lineage}
\alias{functional}
\alias{conditions}
\alias{sample_ids}
\alias{cluster_ids}
\alias{merging_codes}
\alias{merging_ids}
\alias{lineage,daFrame-method}
\alias{functional,daFrame-method}
\alias{conditions,daFrame-method}
\alias{sample_ids,daFrame-method}
\alias{cluster_codes,daFrame-method}
\alias{cluster_ids,daFrame-method}
\title{Extraction and replacement methods for objects of class \code{daFrame}}
\usage{
lineage(x)

functional(x)

sample_ids(x)

cluster_codes(x)

cluster_ids(x)

\S4method{exprs}{daFrame}(object)

\S4method{lineage}{daFrame}(x)

\S4method{functional}{daFrame}(x)

\S4method{conditions}{daFrame}(object)

\S4method{sample_ids}{daFrame}(x)

\S4method{cluster_codes}{daFrame}(x)

\S4method{cluster_ids}{daFrame}(x)
}
\arguments{
\item{x, object}{a \code{\link{daFrame}}.}
}
\value{

}
\description{
Methods for replacing and accessing slots in a \code{\link{daFrame}}.
}
\examples{

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
