% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/daFrame-methods.R
\docType{methods}
\name{n_cells}
\alias{n_cells}
\alias{type_markers}
\alias{state_markers}
\alias{sample_ids}
\alias{cluster_codes}
\alias{cluster_ids}
\alias{exprs,daFrame-method}
\alias{daFrame-methods}
\alias{exprs}
\alias{type_markers}
\alias{state_markers}
\alias{n_cells}
\alias{cluster_codes}
\alias{cluster_ids}
\alias{n_cells,daFrame-method}
\alias{type_markers,daFrame-method}
\alias{state_markers,daFrame-method}
\alias{sample_ids,daFrame-method}
\alias{cluster_codes,daFrame-method}
\alias{cluster_ids,daFrame-method}
\title{Extraction and replacement methods 
for objects of class \code{daFrame}}
\usage{
n_cells(x)

type_markers(x)

state_markers(x)

sample_ids(x)

cluster_codes(x)

cluster_ids(x)

\S4method{exprs}{daFrame}(object)

\S4method{n_cells}{daFrame}(x)

\S4method{type_markers}{daFrame}(x)

\S4method{state_markers}{daFrame}(x)

\S4method{sample_ids}{daFrame}(x)

\S4method{cluster_codes}{daFrame}(x)

\S4method{cluster_ids}{daFrame}(x)
}
\arguments{
\item{x, object}{a \code{\link{daFrame}}.}
}
\value{
\describe{
\item{\code{exprs}}{
extracts the arcsinh-transformed expressions.}
\item{\code{n_cells}}{
extracts the number of events measured per sample.}
\item{\code{type_markers}}{
extracts the antigens used for clustering.}
\item{\code{state_markers}}{
extracts antigens that were not used for clustering.}
\item{\code{sample_ids}}{
extracts the sample IDs as specified in the metadata-table.}
\item{\code{cluster_codes}}{
extracts a \code{data.frame} containing cluster codes for the 
\code{\link{FlowSOM}} clustering, the \code{\link{ConsensusClusterPlus}} 
metaclustering, and all mergings done through \code{\link{mergeClusters}}.}
\item{\code{cluster_ids}}{
extracts the numeric vector of cluster IDs 
as inferred by \code{\link{FlowSOM}}.}
}
}
\description{
Methods for accessing slots in a \code{\link{daFrame}}.
}
\examples{
# construct daFrame
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33", 
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

# view data summary
cbind(metadata(re)$experiment_info, cells=n_cells(re))

# access row / cell data
head(rowData(re))
plot(table(cluster_ids(re)))

# access marker information
type_markers(re)
state_markers(re)

# get cluster ID correspondece between 2 clusterings
old_ids <- seq_len(20)
m <- match(old_ids, cluster_codes(re)$`20`)
new_ids <- cluster_codes(re)$`12`[m]
data.frame(old_ids, new_ids)

# plot relative change in area under CDF curve vs. k
metadata(re)$delta_area

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
