% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{dbFrame-class}
\alias{dbFrame}
\alias{dbFrame-class}
\title{Debarcoding frame class}
\description{
Class representing the data returned by and used throughout debarcoding.
}
\details{
\code{show(dbFrame)} will display \itemize{
\item the dimensionality of the measurement data and number of barcodes used
\item current assignments in order of decreasing population size
\item separation cutoff estimates
\item the average and per-population yield achieven upon debarcoding}
}
\section{Slots}{

\describe{
\item{\code{exprs}}{a matrix containing raw intensities of the input flowFrame.}

\item{\code{bc_key}}{binary barcoding scheme with numeric masses as column names 
and samples names as row names OR a numeric vector of barcode masses.}

\item{\code{bc_ids}}{vector of barcode IDs. If a barcoding scheme is supplied, 
the respective binary code's row name, else, the mass of the respective 
barcode channel.}

\item{\code{deltas}}{numeric vector of separations between positive and negative 
barcode populations computed from normalized barcode intensities.}

\item{\code{normed_bcs}}{matrix containing normalized barcode intensities.}

\item{\code{mhl_dists}}{mahalanobis distances.}

\item{\code{sep_cutoffs}}{numeric vector of distance separation cutoffs between positive and negative 
barcode populations above which events will be unassigned.}

\item{\code{mhl_cutoff}}{non-negative numeric value specifying the Mahalanobis distance cutoffs
below which events will be unassigned.}

\item{\code{counts}}{matrix of dimension (# barcodes)x(101) where each row contains the number 
of events within a barcode for which positive and negative populations 
are separated by a distance between in [0,0.01), ..., [0.99,1], respectively.}

\item{\code{yields}}{a matrix of dimension (# barcodes)x(101) where each row contains the 
percentage of events within a barcode that will be obtained after applying
a separation cutoff of 0, 0.01, ..., 1, respectively.}
}}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

