% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dbFrame-methods.R
\docType{methods}
\name{dbFrame-methods}
\alias{dbFrame-methods}
\alias{bc_ids}
\alias{deltas}
\alias{normed_bcs}
\alias{mhl_dists}
\alias{sep_cutoffs}
\alias{mhl_cutoff}
\alias{counts}
\alias{yields}
\alias{dbFrame-methods}
\alias{exprs}
\alias{bc_key}
\alias{bc_ids}
\alias{deltas}
\alias{normed_bcs}
\alias{mhl_dists}
\alias{sep_cutoffs}
\alias{sep_cutoffs<-}
\alias{mhl_cutoff}
\alias{mhl_cutoff<-}
\alias{counts}
\alias{yields}
\alias{bc_key,dbFrame-method}
\alias{bc_ids,dbFrame-method}
\alias{deltas,dbFrame-method}
\alias{normed_bcs,dbFrame-method}
\alias{mhl_dists,dbFrame-method}
\alias{sep_cutoffs,dbFrame-method}
\alias{mhl_cutoff,dbFrame-method}
\alias{counts,dbFrame-method}
\alias{yields,dbFrame-method}
\alias{mhl_cutoff<-,dbFrame,numeric-method}
\alias{mhl_cutoff<-,dbFrame,ANY-method}
\alias{sep_cutoffs<-,dbFrame,numeric-method}
\alias{sep_cutoffs<-,dbFrame,ANY-method}
\title{Extraction and replacement methods for objects of class \code{dbFrame}}
\usage{
bc_key(object)

bc_ids(object)

deltas(object)

normed_bcs(object)

mhl_dists(object)

sep_cutoffs(object)

mhl_cutoff(object)

counts(object)

yields(object)

\S4method{exprs}{dbFrame}(object)

\S4method{bc_key}{dbFrame}(object)

\S4method{bc_ids}{dbFrame}(object)

\S4method{deltas}{dbFrame}(object)

\S4method{normed_bcs}{dbFrame}(object)

\S4method{mhl_dists}{dbFrame}(object)

\S4method{sep_cutoffs}{dbFrame}(object)

\S4method{mhl_cutoff}{dbFrame}(object)

\S4method{counts}{dbFrame}(object)

\S4method{yields}{dbFrame}(object)

\S4method{mhl_cutoff}{dbFrame,numeric}(object) <- value

\S4method{mhl_cutoff}{dbFrame,ANY}(object) <- value

\S4method{sep_cutoffs}{dbFrame,numeric}(object) <- value

\S4method{sep_cutoffs}{dbFrame,ANY}(object) <- value
}
\arguments{
\item{object}{a \code{\link{dbFrame}}.}

\item{value}{the replacement value.}
}
\description{
\describe{
\item{\code{exprs}}{extract the raw data intensities.}
\item{\code{bc_key}}{extract the barcoding scheme.}
\item{\code{bc_ids}}{extract currently made event assignments.}
\item{\code{deltas}}{extract barcode separations computed from normalized 
                     intensities. \code{sep_cutoffs} apply to these values 
                     (see \code{\link{applyCutoffs}}).}
\item{\code{normed_bcs}}{extracts normalized barcode intensities 
                         (see \code{\link{assignPrelim}}).}
\item{\code{sep_cutoffs}, \code{sep_cutoffs<-}}{extract or replace separation 
cutoffs. If option \code{sep_cutoffs} is not specified, these will be used by 
\code{\link{applyCutoffs}}. Replacement value must be a non-negative numeric 
with length one or same length as the number of rows in the \code{bc_key}.}
\item{\code{mhl_cutoff}, \code{mhl_cutoff<-}}{extract or replace the 
Mahalanobis distance threshold above which events are to be unassigned.
Replacement value must be a single non-negative and non-zero numeric.}
\item{\code{counts}}{extract the counts matrix (see \code{\link{dbFrame}}).}
\item{\code{yields}}{extract the yields matrix (see \code{\link{dbFrame}}).}
}
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
