% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dbFrame-methods.R
\docType{methods}
\name{bc_key}
\alias{<-}
\alias{bc_ids}
\alias{bc_ids,dbFrame-method}
\alias{bc_ids-method}
\alias{bc_key}
\alias{bc_key,dbFrame-method}
\alias{bc_key-method}
\alias{counts}
\alias{counts,dbFrame-method}
\alias{counts-method}
\alias{dbFrame-method}
\alias{dbFrame-methods}
\alias{deltas}
\alias{deltas,dbFrame-method}
\alias{deltas-method}
\alias{exprs}
\alias{exprs,dbFrame-method}
\alias{exprs-method}
\alias{mhl_cutoff}
\alias{mhl_cutoff,dbFrame-method}
\alias{mhl_cutoff-method}
\alias{mhl_cutoff<-,dbFrame,ANY-method}
\alias{mhl_cutoff<-,dbFrame,numeric-method}
\alias{mhl_dists}
\alias{mhl_dists,dbFrame-method}
\alias{mhl_dists-method}
\alias{normed_bcs}
\alias{normed_bcs,dbFrame-method}
\alias{normed_bcs-method}
\alias{sep_cutoffs}
\alias{sep_cutoffs,dbFrame-method}
\alias{sep_cutoffs-method}
\alias{sep_cutoffs<-,dbFrame,ANY-method}
\alias{sep_cutoffs<-,dbFrame,numeric-method}
\alias{yields}
\alias{yields,dbFrame-method}
\alias{yields-method}
\title{dbFrame-methods}
\usage{
bc_key(object)

bc_ids(object)

deltas(object)

normed_bcs(object)

mhl_dists(object)

sep_cutoffs(object)

mhl_cutoff(object)

counts(object)

yields(object)

\S4method{exprs}{dbFrame}(object)

\S4method{bc_key}{dbFrame}(object)

\S4method{bc_ids}{dbFrame}(object)

\S4method{deltas}{dbFrame}(object)

\S4method{normed_bcs}{dbFrame}(object)

\S4method{mhl_dists}{dbFrame}(object)

\S4method{sep_cutoffs}{dbFrame}(object)

\S4method{mhl_cutoff}{dbFrame}(object)

\S4method{counts}{dbFrame}(object)

\S4method{yields}{dbFrame}(object)

\S4method{mhl_cutoff}{dbFrame,numeric}(object) <- value

\S4method{mhl_cutoff}{dbFrame,ANY}(object) <- value

\S4method{sep_cutoffs}{dbFrame,numeric}(object) <- value

\S4method{sep_cutoffs}{dbFrame,ANY}(object) <- value
}

