% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/differentialAbundance.R
\docType{methods}
\name{differentialAbundance}
\alias{differentialAbundance}
\alias{differentialAbundance,flowSet,data.frame,numeric,matrix-method}
\title{differentialAbundance}
\usage{
differentialAbundance(fs, md, cluster_ids, K, ...)

\S4method{differentialAbundance}{flowSet,data.frame,numeric,matrix}(fs, md,
  cluster_ids, K)
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a data.frame containing the metadata.}

\item{cluster_ids}{a numeric vector.}

\item{K}{contrasts.}

\item{...}{optional arguments.}
}
\value{
Returns a table with frequencies across samples,
non-adjusted and adjuster p-values for each cluster.
}
\description{
Compares the proportions of cell types across experimental conditions 
and aims to highlight populations that are present at different ratios.
}
\details{
The blocking variable is patient ID \eqn{i = 1, ..., n}.
\eqn{j = 1, ..., n_i} indicates the sample ID.

We assume that for a given cell population, cell counts \eqn{Y_{ij}}
follow a binomial distribution \eqn{Y_{ij} ~ Bin(m_{ij}, \pi_{ij})}, 
where \eqn{m_{ij}} is the total number of cells in the sample of
patient i and condition j.

The generalized linear mixed model with observation-level random effect
\eqn{\epsilon_{ij}} is defined as follows:

\deqn{E(Y_{ij} | \beta_0, \beta_1, \epsilon_{ij}) = 
logit^{-1}(\beta_0 + \beta_1 x_{ij} + \epsilon_{ij})}
}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
