% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estCutoffs.R
\docType{methods}
\name{estCutoffs}
\alias{estCutoffs}
\alias{estCutoffs,dbFrame-method}
\title{Estimation of distance separation cutoffs}
\usage{
estCutoffs(x, ...)

\S4method{estCutoffs}{dbFrame}(x, deriv = 5)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{deriv}{a single positive integer. Specifies which derivative of the 
three-parameter log-logistic function fit to use for cutoff estimation 
(see below for more details). 
Generally, a low value will yield more stringent (higher) estimates, 
while a higher value will render more liberal (lower) estimates. 
For very distinct bimodality in the count distribution, 
a higher value is recommendable.}
}
\value{
Will update the \code{sep_cutoffs} slot of the input \code{\link{dbFrame}} 
and return the latter.
}
\description{
For each barcode, estimates a cutoff parameter for the 
distance between positive and negative barcode populations.
}
\details{
For the estimation of sample-specific cutoff parameters, we fit a 
three-parameter log-logistic function to the yields function. 
As an adequate cutoff estimate, we target a point which approximately marks 
the end of the plateau regime and on-set of yield decline. 
By default, we compute this as the first minimum of the fifth derivative 
(\code{deriv=5}). However, depending on the overall doublet-singlet 
separation, and how bimodel the count distribution is, another derivate 
may provide a better estimate. 
As a general guideline, higher values of \code{deriv} will shift the 
computed minimum towards the left and yield more liberal (low) cutoffs, 
while a low \code{deriv} will shift it towards the inflection point of 
the yields functions, arriving at more stringent (high) estimates.
}
\examples{
data(sample_ff, sample_key)
# assign preliminary IDs
re <- assignPrelim(x = sample_ff, y = sample_key)
# estimate separation cutoffs
re_d4 <- estCutoffs(x = re, deriv = 4)
re_d5 <- estCutoffs(x = re, deriv = 5)
# view exemplary estimates
plotYields(re_d4, "A1")
plotYields(re_d5, "A1")

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
