% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estCutoffs.R
\docType{methods}
\name{estCutoffs}
\alias{estCutoffs}
\alias{estCutoffs,dbFrame-method}
\title{Estimation of distance separation cutoffs}
\usage{
estCutoffs(x, ...)

\S4method{estCutoffs}{dbFrame}(x, verbose = TRUE)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{verbose}{logical. Should extra information on progress be reported? Defaults to TRUE.}
}
\value{
Will update the \code{sep_cutoffs}, \code{mhl_cutoff}, \code{counts} and 
\code{yields} slots of the input \code{\link{dbFrame}} and return the latter.
}
\description{
For each barcode, estimates a cutoff parameter for the 
distance between positive and negative barcode populations.
}
\examples{
data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)
estCutoffs(x = re)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
