% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estTrim.R
\docType{methods}
\name{estTrim}
\alias{estTrim}
\alias{estTrim,dbFrame-method}
\title{Estimation of optimal trim value}
\usage{
estTrim(x, ...)

\S4method{estTrim}{dbFrame}(x, min = 0.05, max = 0.2, step = 0.01,
<<<<<<< HEAD
  out_path = NULL, name_ext = NULL)
=======
  method = "default", interactions = "default", out_path = NULL,
  name_ext = NULL)
>>>>>>> master
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{min, max, step}{specifies sequence of trim values for which 
compensation should be evaluated.}

<<<<<<< HEAD
=======
\item{method}{function to be used for computing spillover estimates. 
Valid options are \code{"default"} or \code{"classic"} 
(see \code{\link{computeSpillmat}} for details)}

\item{interactions}{\code{"default"} or \code{"all"}. Specifies which interactions spillover 
should be estimated for. The default exclusively takes into consideration 
interactions that are sensible from a chemical and physical point of view
(see \code{\link{computeSpillmat}} for more detail).}

>>>>>>> master
\item{out_path}{specifies in which location output plot is to be generated. 
Defaults to NULL.}

\item{name_ext}{a character string. If specified, will be appended 
to the output plot's name. Defaults to NULL.}
}
\value{
For each value along \code{seq(min, max, step)}, \code{estTrim} will call
\code{\link{computeSpillmat}} with \code{method = "mean"} and the respective 
trim parameter. Returned will be an interactive plot displaying channel-wise 
medians upon compensation, and the mean squared deviation from 0. Each point 
is labeled with the respective interacting channels.
}
\description{
Estimates a trim value that will minimize the sum over squared
popultion- and channel-wise squared medians upon compensation.
}
\examples{
# get single-stained control samples
data(ss_exp)

# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)

re <- assignPrelim(x = ss_exp, y = bc_ms)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
estTrim(x = re, min = 0.02, max = 0.14, step = 0.02)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
