% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{filter,daFrame,character-method}
\alias{filter,daFrame,character-method}
\title{Filter daFrame}
\usage{
\S4method{filter}{daFrame,character}(x, ..., k = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{conditional statements separated by comma.
Only rows/columns where the condition evaluates to TRUE are kept.}

\item{k}{numeric or character string. Specifies the clustering to extract 
populations from. Must be one of \code{names(cluster_codes(x))}.
Defaults to \code{"som100"}.}
}
\value{
a \code{daFrame}.
}
\description{
Filters events/genes from a \code{daFrame} using conditional statements.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
daf <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
daf <- cluster(daf)

# one condition only, remove a single sample
filter(daf, condition == "Ref", sample_id != "Ref1")

# keep only a subset of clusters
filter(daf, cluster_id \%in\% c(7, 8, 18), k = "meta20")

}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
