% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/filter.R
\docType{methods}
\name{filter}
\alias{filter}
\alias{filter,daFrame-method}
\title{Filter daFrame}
\usage{
filter(x, ...)

\S4method{filter}{daFrame}(x, ..., k = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{conditional statements separated by comma.
Left-hand side arguments must occur in the \code{colnames(rowData(x))};
accepted operators are \code{==}, \code{!=} and \code{\%in\%}.}

\item{k}{numeric or character string. 
Specifies the clustering to extract populations from.
Must be one of \code{names(cluster_codes(x))}.}
}
\value{
a \code{daFrame}.
}
\description{
Filters events from a \code{daFrame} using conditional statements.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
daf <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# run clustering
daf <- cluster(daf)

# one condition only, remove a single sample
filter(daf, condition == "Ref", sample_id != "Ref1")

# keep only a subset of clusters
filter(daf, cluster_id \%in\% c(7, 8, 18), k = "meta20")
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
