% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCytof.R
\name{gateCytof}
\alias{gateCytof}
\title{Gating on CyTOF data}
\usage{
gateCytof(
  x,
  chs,
  xy = NULL,
  q = 0.99,
  k = NULL,
  bs = c(1, 0.5),
  type = c("rect", "quad", "elip", "live"),
  gate_id = NULL,
  group_by = NULL,
  gate_on = NULL,
  assay = "exprs",
  overwrite = FALSE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{chs}{character string of length 2 specifying the channels to gate on. 
Should be one of \code{rownames(x)} or a \code{[int_]colData(x)} column,
and correspond to a numeric variable.}

\item{q}{numeric in (0,1) giving the quantile for eliptical and live-gating.}

\item{k}{integer number of cluster centers for eliptical gating.}

\item{bs}{numeric of length 2 specifying the intercept & slope
of the line used for live-gating.}

\item{gate_id}{character string giving a unique identifier for the gate.}

\item{group_by}{character string specifying the grouping variable
in case gates should be applied to specified cell subsets (e.g., 
indenpendently for each sample). Should be a \code{[int_]colData(x)} 
column, and correspond to a logical or factor variable.}

\item{gate_on}{character string specifying a previously applied gate
to gate on, i.e., from which to subset cells prior to gating.
Should be one of \code{names(int_metadata(x)$gates)}.}

\item{assay}{character string specifying which assay data to use.
Should be one of \code{assayNames(x)}.}

\item{overwrite}{logical. If there already exists a gate 
with identifier \code{gate_id}, should it be overwritten?}
}
\value{
a \code{SingleCellExperiment}.
}
\description{
Wrapper around dimension reduction methods available 
through \code{scater}, with optional cell subsampling.
}
\examples{
data(raw_data)
sce <- fcs2sce(raw_data)

# GATING FOR CELLS

# specify DNA channels
dna_chs <- c("Ir191Di", "Ir193Di")

# apply eliptical gate
sce <- gateCytof(sce, dna_chs, 
  type = "elip", q = 0.95, xy = c(4, 4), 
  gate_id = "cells")
  
# view number & fraction of selected cells
table(sce$cells)
mean(sce$cells)

# visualize gate on scatter plot
plotScatter(sce, gate_id = "cells")

}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
