% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessPanel.R
\name{guessPanel}
\alias{guessPanel}
\title{Helper function to parse information from the @parameters slot of a flowFrame/flowSet object}
\usage{
guessPanel(ff, index = 1)
}
\arguments{
\item{ff}{a flowSet or flowFrame object}

\item{index}{if ff is a flowSet object, the index is used to extract a flowFrame}
}
\description{
Helper function to parse information from the @parameters slot of a flowFrame/flowSet object
}
