% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/guessPanel.R
\docType{methods}
\name{guessPanel}
\alias{guessPanel}
\alias{guessPanel,flowFrame-method}
\alias{guessPanel,flowSet-method}
\title{Guess parameter panel}
\usage{
guessPanel(x, ...)

\S4method{guessPanel}{flowFrame}(x)

\S4method{guessPanel}{flowSet}(x, index = 1)
}
\arguments{
\item{x}{a \code{\link{flowFrame}} or \code{\link{flowSet}} object.}

\item{...}{optional arguments.}

\item{index}{if \code{x} is a \code{flowSet} object, 
this index specifies which \code{flowFrame} to extract.}
}
\value{
a \code{data.frame} containing
\itemize{
\item{\code{name}: the parameter name 
as extracted from the input \code{flowFrame},}
\item{\code{desc}: the parameter description 
as extracted from the input \code{flowFrame},}
\item{\code{antigen}: the targeted protein markers, and}
\item{\code{use_channel}: logical. If TRUE, the channel 
is expected to contain a marker and will be kept.}}
}
\description{
Helper function to parse information from the 
\code{parameters} slot of a \code{flowFrame}/\code{flowSet}.
}
\examples{

}
\author{
Mark D Robinson, 
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
