% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normCytof.R
\name{normCytof}
\alias{normCytof}
\title{Bead-based normalization}
\usage{
normCytof(
  x,
  beads,
  assay = "exprs",
  k = 500,
  trim = 5,
  remove_beads = TRUE,
  out_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{beads}{\code{"dvs"} (for bead masses 140, 151, 153 ,165, 175) or 
\code{"beta"} (for bead masses 139, 141, 159, 169, 175) or 
a numeric vector of masses.}

\item{k}{integer width of the median window used for bead smoothing.}

\item{trim}{a single non-negative numeric. 
A \emph{median}+/-\code{trim}*\emph{mad} rule is applied to 
preliminary bead populations to remove bead-bead doublets and 
low signal beads prior to estimating normalization factors.}

\item{remove_beads}{logical. If TRUE, bead events will be removed from
the input \code{SingleCellExperiment} and returned as a separate object?}

\item{out_path}{character string. If specified, outputs will be generated here. 
If NULL (the default), \code{normCytof} will return a \code{SingleCellExperiment} 
with an additional \code{assay} slot \code{"normed"} containing normalized data.
If \code{remove_beads = TRUE}, bead events will be removed and returned separately.}

\item{verbose}{logical. Should extra information on progress be reported?}

\item{fn}{a character string to use as the output file name. Defaults to 
the file name of the input FCS file or \code{flowFrame}, respectively.}

\item{fn_sep}{a character string to use to separate the output file name's prefix
from the appendage.}

\item{norm_to}{a \code{\link{flowFrame}} or character of an FCS file from which baseline 
values should be computed and to which the input data should be normalized.}

\item{plot}{logical. Should bead vs. DNA scatters and smoothed 
bead intensities before vs. after normalization be plotted?}
}
\value{
a list containing with the following elements: 
\itemize{
\item{a \code{SingleCellExperiment} with an additional assay slot \code{normed}}
\item{a \code{ggplot} of DNA vs. bead scatter plots}
\item{a \code{ggplot} of smoothed bead intensities before and after normalization}
}
}
\description{
an implementation of Finck et al.'s normalization of mass cytometry data 
using bead standards with automated bead gating.
}
\examples{
data(raw_data)
sce <- fcs2sce(raw_data)
sce <- transform(sce)
res <- normCytof(sce, beads = "dvs", k = 80)

}
\references{
Finck, R. et al. (2013). 
Normalization of mass cytometry data with bead standards.
\emph{Cytometry A} \bold{83A}, 483-494.
}
\author{
Helena L. Crowell
}
