% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/normCytof.R
\docType{methods}
\name{normCytof}
\alias{normCytof}
\alias{normCytof,character-method}
\title{Bead-based normalization}
\usage{
normCytof(x, y, ...)

\S4method{normCytof}{character}(x, y, gate = "auto", out_path = NULL,
  remove_beads = TRUE, norm_to = NULL)
}
\arguments{
\item{x}{a character string of the FCS file to be normalized.}

\item{y}{\code{"dvs"} or \code{"beta"} or a numeric vector of bead masses.}

\item{...}{optional arguments.}

\item{gate}{\code{"auto"} or \code{"manual"}. 
Specifies how beads should be gated.}

\item{out_path}{a character string. If specified, outputs will be generated in this location. 
If NULL (the default), \code{normCytof} will return a \code{\link{flowFrame}}
of the normalized data (if \code{remove=FALSE}) or a \code{\link{flowSet}} 
containing normalized cells and beads (if \code{remove=TRUE}).}

\item{remove_beads}{logical. If TRUE (the default) normalized cells and beads 
will be returned separately.}

\item{norm_to}{a character string of an FCS file from which baseline values
should be computed and to which the input data should be normalized.}
}
\value{
if \code{out_path=NULL} (the default) a \code{\link{flowFrame}} of the 
normalized data (if \code{remove=FALSE}) or a \code{\link{flowSet}} containing 
normalized cells and beads (if \code{remove=TRUE}). Else, a character of the
location were output FCS files and plots have been generated.
}
\description{
an implementation of Finck et al.'s normalization of mass cytometry data 
using bead standards with automated bead gating.
}
\examples{

}
\references{
Finck, R. et al. (2013).
Normalization of mass cytometry data with bead standards.
\emph{Cytometry A} \bold{83A}, 483–494.
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
