% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/outFCS.R
\docType{methods}
\name{outFCS}
\alias{outFCS}
\alias{outFCS,dbFrame-method}
\title{Write population-wise FCS}
\usage{
outFCS(x, ...)

\S4method{outFCS}{dbFrame}(x, out_path, names = NULL)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{further optional arguments.}

\item{out_path}{character string. Specifies in which location output files are to be generated.}

\item{names}{optional vector of strings specifying output file names. 
By default, each FCS will be given the assigned ID as name.}
}
\description{
Writes an FCS file for each barcode population.
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_beads, y = bc_ms)
#outFCS(x = re, out_path = ...)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}

