% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbMDS.R
\name{pbMDS}
\alias{pbMDS}
\title{Pseudobulk-level MDS plot}
\usage{
pbMDS(
  x,
  color_by = "condition",
  label_by = "sample_id",
  shape_by = NULL,
  assay = "exprs",
  fun = c("median", "mean", "sum")
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{color_by}{character string specifying a 
non-numeric cell metadata column to color by; 
valid values are \code{names(colData(x))}.}

\item{label_by}{character string specifying a 
non-numeric cell metadata column to label by; 
valid values are \code{names(colData(x))}.}

\item{shape_by}{character string specifying a 
non-numeric cell metadata column to shape by; 
valid values are \code{names(colData(x))}.}

\item{assay}{character string specifying which assay data to use;
valid values are \code{assayNames(x)}.}

\item{fun}{character string specifying 
which function to use as summary statistic.}
}
\value{
a \code{ggplot} object.
}
\description{
Pseudobulk-level Multi-Dimensional Scaling (MDS) 
plot computed on median marker expressions in each sample.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
pbMDS(sce)

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
