% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotAbundances.R
\docType{methods}
\name{plotAbundances}
\alias{plotAbundances}
\alias{plotAbundances,daFrame-method}
\title{Population frequencies across samples & clusters}
\usage{
plotAbundances(x, ...)

\S4method{plotAbundances}{daFrame}(x, k = "meta20", by = c("sample_id",
  "cluster_id"), group_by = NULL, shape_by = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{k}{character string. Specifies which clustering to use.}

\item{by}{a character string specifying whether to plot 
frequencies by samples or clusters.}

\item{group_by}{a character string. Should corresponds to a column name of 
\code{rowData(x)} other than "sample_id" and "cluster_id". 
The default NULL will use the first factor available.}

\item{shape_by}{a character string. Should correspond to a column name of
\code{rowData(x)} other than "sample_id" and "cluster_id".}
}
\value{
a \code{ggplot} object.
}
\description{
Plots the relative population abundances of the specified clustering.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
# run clustering
re <- cluster(re)
# plot relative population abundances 
plotAbundances(re, k="meta12")                 # ...by sample 
plotAbundances(re, k="meta8", by="cluster_id") # ...by cluster

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
