% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterExprs.R
\name{plotClusterExprs}
\alias{plotClusterExprs}
\title{Plot expression distributions by cluster}
\usage{
plotClusterExprs(x, k = "meta20", features = NULL)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{k}{character string. Specifies the clustering to use.}

\item{features}{character string specifying which features to include. Defaults to NULL 
(= all features). Alternatively, if the \code{colData(x)$marker_class} 
column is specified, can be one of "type", "state", or "none".}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plots smoothed densities of arcsinh-transformed 
marker intensities by cluster.
}
\examples{
# construct SCE & run clustering
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
sce <- cluster(sce)

plotClusterExprs(sce)

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
