% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterHeatmap.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Plot cluster heatmap}
\usage{
plotClusterHeatmap(
  x,
  hm2 = NULL,
  k = "meta20",
  m = NULL,
  assay = "exprs",
  fun = c("median", "mean", "sum"),
  row_anno = TRUE,
  split_by = NULL,
  scale = TRUE,
  row_dend = TRUE,
  col_dend = FALSE,
  draw_freqs = FALSE,
  hm1_pal = rev(brewer.pal(11, "RdBu")),
  hm2_pal = (if (!is.null(hm2)) if (isTRUE(hm2 == "abundances")) {    
    rev(brewer.pal(11, "PuOr")) } else hm1_pal),
  k_pal = CATALYST:::.cluster_cols,
  m_pal = k_pal
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{hm2}{character string. Specifies the right-hand side heatmap. 
One of: \itemize{
\item{\code{"abundances"}: cluster frequencies across samples}
\item{\code{"state"}: median state-marker expressions 
  across clusters (analogous to the left-hand side heatmap)}
\item{a character string/vector corresponding to one/multiple marker(s): 
  median marker expressions across samples and clusters}}}

\item{k}{character string specifying the clustering 
across which median marker expressions should be computed;
valid values are \code{names(cluster_codes(x))}.}

\item{m}{character string specifying a metaclustering to include 
as a row annotation (for display only, does not effect computations!);
valid values are \code{names(cluster_codes(x))}.}

\item{assay}{character string specifying which assay data to use;
valid values are \code{assayNames(x)}.}

\item{fun}{character string specifying 
the function to use as summary statistic.}

\item{row_anno}{logical. Should row annotations for each
clustering specified by \code{k} (and \code{m}) be included?}

\item{split_by}{character string specifying a cell metadata variable
to group the data by; valid values are \code{names(colData(x))}. 
If specified, a list of multiple heatmaps will be rendered.}

\item{scale}{logical specifying whether expression values should be scaled
between 0 and 1 using lower (1\%) and upper (99\%) quantiles as boundaries.}

\item{row_dend}{logical. Should a row dendrogram
for the hierarchical clustering of clusters be included?}

\item{col_dend}{logical. Should a column dendrogram
for the hierarchical clustering of markers be included?}

\item{draw_freqs}{logical specifying whether to display
a barplot of cell counts labeled with proportions 
for each cluster as a right-hand side row annotation.}

\item{hm1_pal, hm2_pal}{character vector of colors 
to interpolate for each heatmap.}

\item{k_pal, m_pal}{character vector of colors
to use for cluster and merging row annotations.
If less than \code{nlevels(cluster_ids(x, k/m))} 
values are supplied, colors will be interpolated 
via \code{\link[grDevices]{colorRampPalette}}.}
}
\value{
a \code{\link[ComplexHeatmap]{HeatmapList-class}} object.
}
\description{
Plots heatmaps summarizing a clustering and/or metaclustering of interest.
}
\details{
In its 1st panel, \code{plotClusterHeatmap} will display (scaled) 
type-marker expressions aggregated by cluster (across all samples).
Depending on argument \code{hm2}, the 2nd panel will contain one of:
\describe{
\item{\code{hm2 = "abundances"}}{
  relataive cluster abundances by cluster & sample}
\item{\code{hm2 = "state"}}{
  aggregated (scaled) state-marker expressions by 
  cluster (across all samples; analogous to panel 1)}
\item{\code{hm2 \%in\% rownames(x)}}{
  aggregated (scaled) marker expression by cluster & sample}
}
}
\examples{
# construct SCE & run clustering
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
sce <- cluster(sce)

plotClusterHeatmap(sce, hm2="abundances", draw_freqs = TRUE)
plotClusterHeatmap(sce, hm2="pS6", k="meta12", m="meta8")
plotClusterHeatmap(sce, hm2="state", split_by="condition")

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
