% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterHeatmap.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Plot cluster heatmap}
\usage{
plotClusterHeatmap(x, hm2 = NULL, k = "meta20", m = NULL,
  fun = c("median", "mean"), cluster_anno = TRUE, split_by = NULL,
  scale = TRUE, draw_dend = TRUE, draw_freqs = FALSE,
  palette = rev(brewer.pal(11, "RdYlBu")))
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{hm2}{character string. Specifies the right-hand side heatmap. 
One of: \itemize{
\item{\code{"abundances"}: cluster frequencies across samples}
\item{\code{"state_markers"}: median cell state marker expressions 
  across clusters (analogous to the left-hand side heatmap)}
\item{a character string/vector corresponding to one/multiple marker(s): 
  median marker expressions across samples and clusters}}}

\item{k}{character string. Specifies the clustering 
across which median marker expressions should be computed.}

\item{m}{character string. Specifies the metaclustering to be shown. 
(This is for display only and will not effect any computations!)}

\item{fun}{character string specifying the function to use as summary statistic.}

\item{cluster_anno}{logical. Specifies if clusters should be annotated.}

\item{split_by}{character string. Must corresponds to a column name of \code{rowData(x)}. 
If specified, the data will be subset according to this variable, 
and multiple heatmaps will be drawn.}

\item{scale}{logical. Specifies whether scaled values should be plotted.
(see below for details)}

\item{draw_dend}{logical. Specifies if the row dendrogram should be drawn.}

\item{draw_freqs}{logical. Specifyies whether to display cell counts and proportions.}

\item{palette}{character vector of colors to interpolate.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Plots heatmaps summarizing a clustering and/or metaclustering of interest.
}
\details{
Scaled values corresponds to cofactor arcsinh-transformed 
expression values scaled between 0 and 1 using 1% and 99% percentiles as 
boundaries. Hierarchical clustering is performed on the unscaled data.

In its 1st panel, \code{plotClusterHeatmap} will display
median (scaled, arcsinh-transformed) cell-type marker expressions (across all samples).
Depending on argument \code{hm2}, the 2nd panel will contain one of:
\itemize{
\item{relataive cluster abundances by sample}
\item{median (scaled, arcsinh-transformed) 
  cell-state marker expressions (across all samples)}
\item{median (scaled, arcsinh-transformed) 
  cell-state marker expressions by sample}
}
}
\examples{
# construct SCE & run clustering
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
sce <- cluster(sce)

plotClusterHeatmap(sce, hm2="abundances")
plotClusterHeatmap(sce, hm2="abundances", draw_freqs=TRUE)
plotClusterHeatmap(sce, hm2="state_markers", k="meta16", split_by='condition')
plotClusterHeatmap(sce, hm2="pS6", k="meta12", m="meta8")
plotClusterHeatmap(sce, hm2="abundances", scale=FALSE, draw_freqs=TRUE)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
