% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotClusteringHeatmap.R
\docType{methods}
\name{plotClusteringHeatmap}
\alias{plotClusteringHeatmap}
\alias{plotClusteringHeatmap,matrix,vector-method}
\title{Median marker expressions across clusters}
\usage{
plotClusteringHeatmap(x, cluster_ids, ...)

\S4method{plotClusteringHeatmap}{matrix,vector}(x, cluster_ids,
  merging_ids = NULL, anno = FALSE)
}
\arguments{
\item{...}{optional arguments.}

\item{anno}{logical indicating whether to display values inside each bin}

\item{fs}{}
}
\value{


Heatmap of median marker expressions across clusters.
}
\description{

}
\details{
Displayed values corresponds to cofactor 5 arcsinh-transformed 
expression values scaled between 0 and 1 using 1% and 99% percentiles as 
boundaries. Hierarchical clustering is performed on the unscaled data.
}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
