% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotCounts.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,daFrame-method}
\title{Number of events per sample}
\usage{
plotCounts(x, ...)

\S4method{plotCounts}{daFrame}(x, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{color_by}{a character string that appears as a column name in the
metadata-table of the input \code{daFrame}. Specifies the color coding.
In the case of multiple conditions, \code{"condition"} will color points
according to each sample's combined conditions.}

\item{out_path}{a character string. If specified, 
output will be generated in this location. Defaults to NULL.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Barplot of the number of cells measured for each sample.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotCounts(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
