% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDiffHeatmap.R
\docType{methods}
\name{plotDiffHeatmap}
\alias{plotDiffHeatmap}
\alias{plotDiffHeatmap,daFrame-method}
\title{Median marker expressions across samples}
\usage{
plotDiffHeatmap(x, K, ...)

\S4method{plotDiffHeatmap}{daFrame}(x, K, th = 2.5, FDR_cutoff = 0.05,
  palette = brewer.pal(n = 11, name = "RdYlGn"), out_path = NULL)
}
\arguments{
\item{x}{expression matrix.}

\item{...}{optional arguments.}

\item{palette}{a character string specifying the colors to interpolate.}

\item{out_path}{a character string. If specified, 
output will be generated in this location. Defaults to NULL.}

\item{anno}{logical. Specifies whether to display values insinde each bin.}
}
\value{


an object of class \code{HeatmapList{\link{ComplexHeatmap}}}.
}
\description{
Plots median marker expressions across samples
computed on arcsinh-transformed intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotExprHeatmap(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
