% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDiffHeatmap.R
\docType{methods}
\name{plotDiffHeatmap}
\alias{plotDiffHeatmap}
\alias{plotDiffHeatmap,daFrame,SummarizedExperiment-method}
\title{Plot differential heatmap}
\usage{
plotDiffHeatmap(x, y, ...)

\S4method{plotDiffHeatmap}{daFrame,SummarizedExperiment}(x, y, top_n = 20,
  all = FALSE, order = TRUE, th = 0.1)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{y}{a \code{SummarizedExperiment} containing differential testing
results as returned by one of \code{\link[diffcyt]{testDA_edgeR}}, 
\code{\link[diffcyt]{testDA_voom}}, \code{\link[diffcyt]{testDA_GLMM}}, 
\code{\link[diffcyt]{testDS_limma}}, or \code{\link[diffcyt]{testDS_LMM}}.}

\item{...}{optional arguments.}

\item{top_n}{numeric. Number of top clusters (if \code{type = "DA"}) or
cluster-marker combinations (if \code{type = "DS"}) to display.}

\item{all}{logical. Specifies whether all clusters or cluster-marker combinations 
should be displayed. If \code{TRUE}, \code{top_n} will be ignored.}

\item{order}{logical. Should results be ordered by significance?}

\item{th}{numeric. Threshold on adjusted p-values below which clusters (DA) 
or cluster-marker combinations (DS) should be considered significant.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Heatmaps summarizing differental abundance 
& differential state testing results.
}
\details{
For DA tests, \code{plotDiffHeatmap} will display
\itemize{
\item median (arcsinh-transformed) cell-type marker expressions (across all samples)
\item cluster abundances by samples
\item row annotations indicating the significance of detected clusters
}
For DS tests, \code{plotDiffHeatmap} will display
\itemize{
\item median (arcsinh-transformed) cell-type marker expressions (across all samples)
\item median (arcsinh-transformed) cell-state marker expressions (across all samples)
\item median (arcsinh-transformed) cell-state marker expressions by sample
\item row annotations indicating the significance of detected cluster-marker combinations
}
}
\author{
Lukas M Weber, Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
