% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDR.R
\name{plotDR}
\alias{plotDR}
\title{Plot reduced dimensions}
\usage{
plotDR(x, dr = NULL, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{dr}{character string specifying which dimension reduction to use. 
Should be one of \code{reducedDimNames(x)}; default to the 1st available.}

\item{color_by}{character string corresponding to a
\code{colData(x)} column. Specifies the color coding.}
}
\value{
a \code{ggplot} object.
}
\description{
Dimension reduction plot colored 
by expression, cluster, sample or group ID.
}
\examples{
# construct SCE & run clustering
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
sce <- cluster(sce)

library(scater)
sce <- runUMAP(sce, exprs_values = "exprs")

# color by pS6 expression, split by group
plotDR(sce, color_by = "pS6") +
  facet_wrap("condition")

# color by 8 metaclusters, split by sample
plotDR(sce, color_by = "meta8") + 
  facet_wrap("sample_id", ncol = 4)

}
\references{
Nowicka M, Krieg C, Crowell HL, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
