% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEvents.R
\docType{methods}
\name{plotEvents}
\alias{plotEvents}
\alias{plotEvents,dbFrame-method}
\title{Event plot}
\usage{
plotEvents(x, ...)

\S4method{plotEvents}{dbFrame}(x, which_bc = "all", n_events = 100,
  out_path = NULL, name_ext = NULL)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{further optional arguments.}

\item{which_bc}{numeric or "all". Specifies which barcode to plot. Defaults to "all".}

\item{n_events}{numeric or "all". Specifies how many events to plot per barcode.}

\item{out_path}{a character string. If specified, outputs will be generated in this location. 
Defaults to NULL.}

\item{name_ext}{a character string. If specified, will be appended to the plot's name. 
Defaults to NULL.}
}
\description{
Shows normalized barcode intensities for a given barcode.
}
\details{
EXPLAIN NORMALIZATION HERE
}
\examples{
data(ss_beads)
bc_ms <- c(139, 141:156, 158:176)
re <- assignPrelim(x = ss_beads, y = bc_ms)
plotEvents(x = re, which_bc = 139)

}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}

