% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEvents.R
\docType{methods}
\name{plotEvents}
\alias{plotEvents}
\alias{plotEvents,dbFrame-method}
\title{Event plot}
\usage{
plotEvents(x, ...)

\S4method{plotEvents}{dbFrame}(x, which = "all", n_events = 100,
  out_path = NULL, name_ext = NULL)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{which}{"all", numeric or character. Specifies which barcode(s) to plot. 
Valid values are IDs that occur as row names in the \code{bc_key} 
of the supplied \code{\link{dbFrame}}, or 0 for unassigned events. 
Defaults to "all".}

\item{n_events}{numeric. Specifies number of events to plot. Defaults to 100.}

\item{out_path}{a character string. If specified, outputs will be generated 
in this location. Defaults to NULL.}

\item{name_ext}{a character string. If specified, will be appended to the plot's name. 
Defaults to NULL.}
}
\value{
plots intensities normalized by population for each barcode specified
by \code{which}: Each event corresponds to the intensities plotted on a 
vertical line at a given point along the x-axis. Events are scaled to the 
95\% quantile of the population it has been assigned to. Barcodes with 
less than 50 event assignments will be skipped; it is strongly recoomended
to remove such populations or reconsider their separation cutoffs.
}
\description{
Shows normalized barcode intensities for a given barcode.
}
\examples{
data(sample_ff, sample_key)

# view preliminary assignments
re <- assignPrelim(x = sample_ff, y = sample_key)
plotEvents(x = re, which = "D1", n_events = 1000)

# apply deconvolution parameters
re <- estCutoffs(re)
re <- applyCutoffs(x = re)
plotEvents(x = re, which = "D1", n_events = 500)

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
