% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExprHeatmap.R
\docType{methods}
\name{plotExprHeatmap}
\alias{plotExprHeatmap}
\alias{plotExprHeatmap,daFrame-method}
\title{Median marker expressions across samples}
\usage{
plotExprHeatmap(x, ...)

\S4method{plotExprHeatmap}{daFrame}(x, anno = TRUE, color_by = "condition",
  palette = brewer.pal(n = 8, name = "YlGnBu"), out_path = NULL)
}
\arguments{
\item{x}{expression matrix.}

\item{...}{optional arguments.}

\item{anno}{logical. Specifies whether to display values insinde each bin.}

\item{color_by}{a character string that specifies the row annotation.}

\item{palette}{a character string specifying the colors to interpolate.}

\item{out_path}{a character string. If specified, 
output will be generated in this location. Defaults to NULL.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Plots median marker expressions across samples
computed on arcsinh-transformed intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotExprHeatmap(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
