% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExprHeatmap.R
\docType{methods}
\name{plotExprHeatmap}
\alias{plotExprHeatmap}
\alias{plotExprHeatmap,daFrame-method}
\title{Plot expression heatmap}
\usage{
plotExprHeatmap(x, ...)

\S4method{plotExprHeatmap}{daFrame}(x, anno = TRUE, color_by = "condition",
  palette = brewer.pal(n = 8, name = "YlGnBu"), scale = TRUE,
  draw_freqs = FALSE, clustering_distance = "euclidean",
  clustering_linkage = "average")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{anno}{logical. Specifies whether to display values insinde each bin.}

\item{color_by}{character string. Specifies the row annotation.}

\item{palette}{character vector of colors to interpolate.}

\item{scale}{logical. Specifies whether to mean-variance normalize each channel.}

\item{draw_freqs}{logical. Specifyies whether to display cell counts and proportions.}

\item{clustering_distance}{a character string that specifies the metric to use in 
\code{\link[stats]{dist}()} for clustering.}

\item{clustering_linkage}{a character string that specifies the linkage to use in
\code{\link[stats]{hclust}()} for clustering.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Plots median marker expressions across samples 
computed on arcsinh-transformed intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotExprHeatmap(re)
plotExprHeatmap(re, scale=FALSE)
plotExprHeatmap(re, plot_freqs=TRUE)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
