% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExprHeatmap.R
\name{plotExprHeatmap}
\alias{plotExprHeatmap}
\title{Plot expression heatmap}
\usage{
plotExprHeatmap(
  x,
  bin_anno = TRUE,
  row_anno = TRUE,
  palette = brewer.pal(n = 8, name = "YlGnBu"),
  scale = TRUE,
  draw_freqs = FALSE,
  clustering_distance = "euclidean",
  clustering_linkage = "average"
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{bin_anno}{logical. Specifies whether to display values inside bins.}

\item{row_anno}{logical. Should row annotations for each factor 
in \code{metadata(x)$experiment_info} be included?}

\item{palette}{character vector of colors to interpolate.}

\item{scale}{logical. Should scaled values be displayed? (see details)}

\item{draw_freqs}{logical. Should cell counts and proportions be displayed?}

\item{clustering_distance}{character string that specifies 
the metric to use in \code{\link[stats]{dist}} for clustering.}

\item{clustering_linkage}{character string that specifies 
the linkage to use in \code{\link[stats]{hclust}} for clustering.}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Plots median marker expressions across samples 
computed on arcsinh-transformed intensities.
}
\details{
Scaled values corresponds to cofactor arcsinh-transformed 
expression values scaled between 0 and 1 using 1% and 99% percentiles as 
boundaries. Hierarchical clustering is performed on the unscaled data.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)
plotExprHeatmap(sce, draw_freqs=TRUE)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
