% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExprs.R
\docType{methods}
\name{plotExprs}
\alias{plotExprs}
\alias{plotExprs,daFrame-method}
\title{Plot expressions}
\usage{
plotExprs(x, ...)

\S4method{plotExprs}{daFrame}(x, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{color_by}{character string. Has to appear as a column name of \code{rowData(x)}.
Specifies the color coding.}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plots the smoothed densities of arcsinh-transformed marker intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
plotExprs(re)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
