% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExprs.R
\docType{methods}
\name{plotExprs}
\alias{plotExprs}
\alias{plotExprs,flowSet,data.frame-method}
\title{Smoothed densities of marker expression}
\usage{
plotExprs(fs, md, ...)

\S4method{plotExprs}{flowSet,data.frame}(fs, panel, md, cols = c("steelblue",
  "indianred"))
}
\arguments{
\item{fs}{a \code{\link{flowSet}} holding all samples.}

\item{md}{a \code{data.frame} containing the following columns:
\itemize{
\item \code{file_name}: names of the fcs files contained in the input flowSet
\item \code{sample_id}: a unique identifier for each sample
\item \code{condition}: sample description (e.g. healthy or diseased)
}}

\item{...}{optional arguments.}
}
\value{
Plots sample-wise smoothed densities of marker expressions.
}
\description{
Sample-wise smoothed densities of marker expressions.
}
\examples{
data(sample_ff, sample_key)
assignPrelim(x = sample_ff, y = sample_key)

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
