% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMahal.R
\docType{methods}
\name{plotMahal}
\alias{plotMahal}
\alias{plotMahal,dbFrame-method}
\title{Biaxial plot}
\usage{
plotMahal(x, ...)

\S4method{plotMahal}{dbFrame}(x, which, cofactor = 50, out_path = NULL,
  name_ext = NULL)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{which}{specifies which barcode to plot.}

\item{cofactor}{cofactor used for asinh transformation.}

\item{out_path}{a character string. If specified, outputs will be generated 
in this location. Defaults to NULL.}

\item{name_ext}{a character string. If specified, will be appended to the plot's name. 
Defaults to NULL.}
}
\value{
plots all inter-barcode interactions for the population specified
by argument \code{which}. Events are colored by their Mahalanobis distance.
}
\description{
Histogram of counts and plot of yields as a function of separation cutoffs.
}
\examples{
data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)
re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
plotMahal(x = re, which = "B3")

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
