% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMedExprs.R
\docType{methods}
\name{plotMedExprs}
\alias{plotMedExprs}
\alias{plotMedExprs,daFrame-method}
\title{Median marker expressions across samples}
\usage{
plotMedExprs(x, ...)

\S4method{plotMedExprs}{daFrame}(x, k = 20, facette = c("antigen",
  "cluster_id"), group_by = "condition")
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{k}{a numeric or charactering specifying the clustering to use.
If \code{facette = "antigen"} this argument will be ignored.}

\item{facette}{character string \code{"antigen"} or \code{"cluster_id"}. 
Note that the latter requires having run \code{\link{cluster}} first.}

\item{group_by}{a character string that appears as a column name in the 
metadata-table of the input \code{daFrame}. Specifies sample grouping.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots median marker expressions across samples
computed on arcsinh-transformed intensities.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

# plot median expressions
plotMedExprs(re)

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33", 
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

# plot median expressions across clusters
plotMedExprs(re, facette="cluster_id", k=8)

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
