% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMedExprs.R
\name{plotMedExprs}
\alias{plotMedExprs}
\title{Plot median expressions}
\usage{
plotMedExprs(
  x,
  k = "meta20",
  facet = c("antigen", "cluster_id"),
  group_by = "condition",
  shape_by = NULL
)
}
\arguments{
\item{x}{a \code{\link{SingleCellExperiment}{SingleCellExperiment}}.}

\item{k}{character string. Specifies the clustering to use.
If \code{facet = "antigen"}, this argument will be ignored.}

\item{facet}{\code{"antigen"} or \code{"cluster_id"}. 
Note that the latter requires having run \code{\link{cluster}}.}

\item{group_by}{character string specifying 
a \code{colData(x)} column to group samples by.}

\item{shape_by}{character string specifying 
a \code{colData(x)} column to shape samples by.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots median marker expressions across samples
computed on arcsinh-transformed intensities.
}
\examples{
# construct SCE
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)

# plot median expressions
plotMedExprs(sce, shape_by = "patient_id")

# run clustering
sce <- cluster(sce)

# plot median expressions across clusters
plotMedExprs(sce, facet = "cluster_id", k = "meta8")

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell
}
