% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNRS.R
\docType{methods}
\name{plotNRS}
\alias{plotNRS}
\alias{plotNRS,flowSet,data.frame,data.frame-method}
\title{Non-redundancy scores across markers}
\usage{
plotNRS(fs, panel, md, ...)

\S4method{plotNRS}{flowSet,data.frame,data.frame}(fs, panel, md,
  cols = c("steelblue", "indianred"))
}
\arguments{
\item{fs}{}

\item{md}{}

\item{...}{optional arguments.}
}
\value{

}
\description{

}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
