% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNRS.R
\name{plotNRS}
\alias{plotNRS}
\title{Plot non-redundancy scores}
\usage{
plotNRS(x, features = NULL, color_by = "condition")
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{features}{character string specifying which features to include. Defaults to NULL 
(= all features). Alternatively, if the \code{rowData(x)$marker_class} 
column is specified, can be one of "type", "state", or "none".}

\item{color_by}{character string. Has to appeara as a column name of \code{colData(x)}.
Specifies the color coding.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots non-redundancy scores (NRS) by features in decreasing order.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
sce <- prepData(PBMC_fs, PBMC_panel, PBMC_md)

plotNRS(sce, features = NULL)   # default: all markers
plotNRS(sce, features = "type") # type-markers only

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{helena.crowell@uzh.ch}
}
