% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotNRS.R
\docType{methods}
\name{plotNRS}
\alias{plotNRS}
\alias{plotNRS,daFrame-method}
\title{Non-redundancy scores across markers}
\usage{
plotNRS(x, ...)

\S4method{plotNRS}{daFrame}(x)
}
\arguments{
\item{x}{a \code{\link{daFrame}}}

\item{...}{optional arguments.}
}
\value{
Plots non-redundancy scores (NRS) across lineage markers.
}
\description{

}
\examples{

}
\references{
Nowicka M, Krieg C, Weber LM et al.
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
