% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Scatter plot}
\usage{
plotScatter(
  x,
  chs = NULL,
  gate_id = NULL,
  show_gate = TRUE,
  show_perc = TRUE,
  bins = 100,
  assay = "exprs",
  color_by = c("density", "selection")
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{chs}{character string of length 2 specifying which channels to plot.
Should be one of \code{rownames(x)} or \code{names([int_]colData(x))},
and correspond to a numeric or logical variable.}

\item{gate_id}{a character vector specifying 
a gate added via \code{\link{gateCytof}}.}

\item{show_gate}{logical. If \code{gate_id} is specified, specifies 
whether the boundary of the specified gate should be visualized.}

\item{show_perc}{logical. If \code{gate_id} is specified, specifies 
whether the percentage of gated cells should be included.}

\item{assay}{character string specifying which assay data to use.
Should be one of \code{assayNames(x)}.}

\item{color_by}{specifies the color coding. If \code{gate_id} is specified, 
\code{color_by = "selection"} will color cells in a binary fashion, 
i.e., whether cells fall within the gate's boundary or not.}
}
\value{
a \code{ggplot} object.
}
\description{

}
\examples{

}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
