% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotSNE.R
\docType{methods}
\name{plotSNE}
\alias{plotSNE}
\alias{plotSNE,daFrame-method}
\title{MDS plot}
\usage{
plotSNE(x, ...)

\S4method{plotSNE}{daFrame}(x, color_by = 20, facet = NULL)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{optional arguments.}

\item{color_by}{numeric value or character string specifying a clustering
OR a character string specifying an antibody whose expression to color by.}

\item{facet}{a character string specifying a factor to subset the data by.
One of \code{names(rowData(x))}. Defaults to NULL.}
}
\value{
a \code{ggplot} object.
}
\description{
Multi-dimensional scaling (MDS) plot on median marker expressions.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
re <- tSNE(re)
plotSNE(re, color_by=12)
plotSNE(re, color_by="pS6", facet="condition")

}
\references{
Nowicka M, Krieg C, Weber LM et al. 
CyTOF workflow: Differential discovery in 
high-throughput high-dimensional cytometry datasets.
\emph{F1000Research} 2017, 6:748 (doi: 10.12688/f1000research.11622.1)
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
