% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpillmat.R
\name{plotSpillmat}
\alias{plotSpillmat}
\title{Spillover matrix heatmap}
\usage{
plotSpillmat(
  x,
  sm = NULL,
  out_path = NULL,
  name_ext = NULL,
  anno = TRUE,
  plotly = FALSE,
  isotope_list = NULL
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{sm}{spillover matrix to visualize. If NULL, \code{plotSpillmat} 
will try and access \code{metadata(x)$spillover_matrix}.}

\item{out_path}{character string. If specified, outputs will be generated here.}

\item{name_ext}{character string. If specified, will be appended to the plot's name.}

\item{anno}{logical. If TRUE (default), spill percentages are shown inside 
bins and rows are annotated with the total amount of spill received.}

\item{plotly}{logical. Should an interactive plot be rendered?}

\item{isotope_list}{named list. Used to validate the input spillover matrix.
Names should be metals; list elements numeric vectors of their isotopes.
See \code{\link{isotope_list}} for the list of isotopes used by default.}
}
\value{
a \code{ggplot2}-object showing estimated spill percentages 
  as a heatmap with colors ramped to the highest spillover value present.
}
\description{
Generates a heatmap of the spillover matrix
annotated with estimated spill percentages.
}
\examples{
# get single-stained control samples & construct SCE
data(ss_exp)
sce <- fcs2sce(ss_exp)

# debarcode single-positive populations
bc_ms <- c(139, 141:156, 158:176)
sce <- assignPrelim(x = sce, bc_key = bc_ms, verbose = FALSE)
sce <- applyCutoffs(estCutoffs(x = sce))

# estimate & visualize spillover matrix
sce <- computeSpillmat(x = sce)
plotSpillmat(sce)

}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
