% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotYields.R
\docType{methods}
\name{plotYields}
\alias{plotYields}
\alias{plotYields,dbFrame-method}
\title{Yield plot}
\usage{
plotYields(x, ...)

\S4method{plotYields}{dbFrame}(x, which = 0, out_path = NULL,
  name_ext = NULL, plotly = TRUE)
}
\arguments{
\item{x}{a \code{\link{dbFrame}}.}

\item{...}{optional arguments.}

\item{which}{0, numeric or character. Specifies which barcode(s) to plot. 
Valid values are IDs that occur as row names of \code{bc_key(x)}; 
0 (the default) will generate a summary plot with all barcodes.}

\item{out_path}{character string. If specified, outputs will be generated here.}

\item{name_ext}{character string. If specified, will be appended to the plot's name.}

\item{plotly}{logical. Should an interactive plot be rendered?}
}
\value{
plots the distribution of barcode separations and yields upon 
debarcoding as a function of separation cutoffs. If available, currently 
used separation cutoffs as well as their resulting yields will be indicated 
in the plot`s main title.
}
\description{
Distribution of barcode separations and 
yields as a function of separation cutoffs.
}
\details{
The overall yield that will be achieved upon application of the specified 
set of separation cutoffs is indicated in the summary plot. Respective 
separation thresholds and their resulting yields are included in each 
barcode's plot. The separation cutoff value should be chosen such that
it appropriately balances confidence in barcode assignment and cell yield.
}
\examples{
data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)
re <- estCutoffs(x = re)

# all barcodes summary plot
plotYields(x = re, which = 0)

# plot for specific sample
plotYields(x = re, which = "C1")

}
\references{
Zunder, E.R. et al. (2015).
Palladium-based mass tag cell barcoding with a doublet-filtering scheme 
and single-cell deconvolution algorithm.
\emph{Nature Protocols} \bold{10}, 316-333.
}
\author{
Helena Lucia Crowell \email{crowellh@student.ethz.ch}
}
