% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Data preparation}
\usage{
prepData(
  x,
  panel,
  md,
  features = NULL,
  cofactor = 5,
  panel_cols = list(channel = "fcs_colname", antigen = "antigen", class =
    "marker_class"),
  md_cols = list(file = "file_name", id = "sample_id", factors = c("condition",
    "patient_id"))
)
}
\arguments{
\item{x}{a \code{flowSet} holding all samples 
or a path to a set of FCS files.}

\item{panel}{a data.frame containing, for each channel, 
its column name in the input data, targeted protein marker,
and (optionally) class ("type", "state", or "none").}

\item{md}{a table with column describing the experiment.
An exemplary metadata table could look as follows: \itemize{
\item\code{file_name}: the FCS file name
\item\code{sample_id}: a unique sample identifier
\item\code{patient_id}: the patient ID
\item\code{condition}: brief sample description 
  (e.g. reference/stimulated, healthy/diseased)}}

\item{features}{a logical vector, numeric vector of column indices,
or character vector of channel names. Specified which column to keep 
from the input data. Defaults to the channels listed in the input panel.}

\item{cofactor}{numeric cofactor to use for arcsinh-transformation.}

\item{panel_cols}{a names list specifying the column names of \code{panel}
that contain the channel names, targeted protein markers, and (optionally) 
marker classes.}

\item{md_cols}{a named list specifying the column names of \code{md}
that contain the FCS file names, sample IDs, and factors of interest
(batch, condition, treatment etc.).}
}
\value{
an object of class \code{daFrame}.
}
\description{
Data preparation
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
prepData(PBMC_fs, PBMC_panel, PBMC_md)

}
\author{
Helena L Crowell \email{helena.crowell@uzh.ch}
}
