% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runDR.R
\docType{methods}
\name{runDR}
\alias{runDR}
\alias{runDR,daFrame-method}
\title{Perform dim. reduction on a \code{daFrame}}
\usage{
runDR(x, ...)

\S4method{runDR}{daFrame}(x, dr = c("TSNE", "PCA", "MDS", "UMAP",
  "DiffusionMap"), rows_to_use = 1000, cols_to_use = NULL,
  overwrite = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{daFrame}}.}

\item{...}{additional parameters passed to the dim. reduction method
(see \code{\link[scater]{runTSNE}}, \code{\link[scater]{runPCA}}, 
\code{\link[scater]{runMDS}}, \code{\link[scater]{runUMAP}}, 
and \code{\link[scater]{runDiffusionMap}}).}

\item{dr}{character string specifying the dimensionaly reduction method.}

\item{rows_to_use}{numeric vector of row indices (cells) to use.
If NULL, all cells will be used. If a single integer value N,
(default 1000) a subset of N cells will be drawn from each sample.}

\item{cols_to_use}{character vector in \code{colnames(x)} or numeric
vector of column indices to use for computing reduced dimensions.}

\item{overwrite}{logical. Whether to force overwriting 
any existing dimension reduction of type \code{dr}.}
}
\value{
a \code{daFrame} with an additional entry titled "\code{dr}"
  in the \code{reducedDims} slot of the input \code{daFrame}.
}
\description{
Wrapper function to perform dimensionality reduction methods
  on a \code{daFrame} object using \code{scater}.
}
\examples{
data(PBMC_fs, PBMC_panel, PBMC_md)
daf <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)
daf <- cluster(daf)

daf <- runDR(daf, "PCA", rows_to_use = cells_use)
plotDR(daf, "PCA", color_by = "condition")

# downsample to 100 cells per sample
cells_by_sample <- split(seq_len(nrow(daf)), sample_ids(daf))
cells_use <- unlist(lapply(cells_by_sample, sample, 100))

daf <- runDR(daf, "TSNE", rows_to_use = cells_use, use_dimred = "PCA")
plotDR(daf, "TSNE", color_by = "meta20")

daf <- runDR(daf, "UMAP", rows_to_use = cells_use)
plotDR(daf, "UMAP", color_by = "meta10")

}
\author{
Helena L. Crowell \email{helena.crowell@uzh.ch}
}
